### 2025-03-13 #7896 @editorialbot
[PRE REVIEW]: pythonradex: a python implementation of the non-LTE radiative transfer code RADEX with additional functionality
**Submitting author:** @gica3618 (<a href="http://orcid.org/0000-0002-2700-9676">Gianni Cataldi</a>)
**Repository:** https://github.com/gica3618/pythonradex
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.0.9
