+++ joss_paper/paper.bib
@@ -15,7 +15,6 @@ archivePrefix = {arXiv},
        url = {https://ui.adsabs.harvard.edu/abs/2007A&A...468..627V}
 }
 
-
 @inbook{Rybicki1985,
 author = {{Rybicki}, G.~B. and {Lightman}, A.~P.},
 doi = {10.1002/9783527618170.ch1},
@@ -25,7 +24,7 @@ title = {FUNDAMENTALS OF RADIATIVE TRANSFER},
 booktitle = {Radiative Processes in Astrophysics},
 chapter = {1},
 pages = {1-50},
-doi = {https://doi.org/10.1002/9783527618170.ch1},
+doi = {10.1002/9783527618170.ch1},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/9783527618170.ch1},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/9783527618170.ch1},
 year = {1985},
