### 2025-03-14 #7896 @gica3618
Suggestions for potential reviewers, based purely on a keyword search in the [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review (I don't know any of these people or their work):

- [nenasedk](https://github.com/nenasedk)
- [arjunsavel](https://github.com/arjunsavel)
- [pscicluna](https://github.com/pscicluna)
- [bronreichardtchu](https://github.com/bronreichardtchu)
