### 2025-03-16 #7896 @warrickball
Hi @gica3618, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience! In the meantime, I'll make some preliminary checks in the next few days and possibly comment again but at a glance everything looks ready for review.
