### 2025-05-28 81ba939
better error message when user requests unavailable collider
