### 2025-05-29 91eae28
better error message when unavailable collider is requested
