### 2025-07-10 65da3e0
moved get_Tex from EmittingMolecule to Molecule
