### 2025-10-13 #9233 @editorialbot
[REVIEW]: pythonradex: a fast Python re-implementation of RADEX with extended functionality
**Submitting author:** @gica3618 (<a href="http://orcid.org/0000-0002-2700-9676">Gianni Cataldi</a>)
**Repository:** https://github.com/gica3618/pythonradex
**Branch with paper.md** (empty if default branch): development
**Version:** v2.0.0
