### 2025-10-30 74dc3af
updates to documenation to mention EMAA database at alternative to LAMDA
