### 2025-10-31 #9233 @andresmegias
> Hi everyone, just checking in here! [@pscicluna](https://github.com/pscicluna) and [@GijsVermarien](https://github.com/GijsVermarien), did you have any initial comments that the authors could start working on? Remember that JOSS reviews are intended to be iterative and not monolithic like traditional journals.
>
> [@andresmegias](https://github.com/andresmegias), let me know if everything makes sense, and don't wait to create your reviewer checklist by commenting here with `@editorialbot generate my checklist`, even if you don't think you have time for a full review just yet.

Hi all,

Thank you for the messages, Warrick. I've read the post and the reviewer instructions, I've created my checklist, and did some of the checks. I'll be able to work more on this next week after Tuesday.
