### 2025-11-05 gica3618/pythonradex#9 @GijsVermarien
Dependency nitpicks
Hi Gianni,

I am currently trying to install pythonradex, I have a few small nitpicks:
- Python 3.14 is not yet supported. Is this just because it was recently released or is there some problem with using the latest Python?
- Could you add matplotlib as a dependency in your `pyproject.toml`? Then you can immediately run all notebooks after install with no further pip commands.

Best,

Gijs
