### 2025-11-05 gica3618/pythonradex#10 @GijsVermarien
Adding a contributing.md
Could you add your contributing information on the Github as well? https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

https://pythonradex.readthedocs.io/en/latest/help.html

Also, is there an email address people can write mails to?
