### 2025-11-05 #9233 @GijsVermarien
@gica3618 I think the paper illustrates your code really well and is tied to some very nice documentation notebooks. I have some questions/suggestions for the paper:

L8 - You mention we can derive physical parameters, I understand that we can derive temperature and density using a radiative transfer code, however mass is not that easy to derive without many more assumptions?
L28 - A line explaining what this iterative solving process looks like for non-LTE systems would be useful.
L50 - Here it would be useful to briefly explain the LVG assumption.
Figure 1 - Could you add the black dotted lines to the legend?
L60 - I would like to see a figure that illustrates this comparison

General remark: you could add the citation for the collisional data files from LAMBDA([10.1051/0004-6361:20041729](https://ui.adsabs.harvard.edu/link_gateway/2005A&A...432..369S/doi:10.1051/0004-6361:20041729)) and EMAA(http://dx.doi.org/10.17178/EMAA) database since they are a crucial part of your workflow.
