### 2025-11-07 f7ed609
added jupyter and matplotlib as optional dependencies to run notebooks; added pytest as optional dependency for dev; updated documentation
