+++ docs/installation.rst
@@ -7,6 +7,12 @@ Installation
 
     pip install pythonradex
 
+If you would like to run the Jupyter notebooks presented in the :ref:`examples` section, you can include the additional dependencies (``jupyter`` and ``matplotlib``) by installing ``pythonradex`` as follows:
+
+.. code-block:: bash
+
+    pip install pythonradex[notebooks]
+
 ``pythonradex`` is also available on `anaconda <https://anaconda.org/conda-forge/pythonradex>`_. Use the following command to install it with conda:
 
 .. code-block:: bash
