+++ pyproject.toml
@@ -23,6 +23,10 @@ classifiers = [
 dependencies = ['scipy>=1.14','numpy>=2.0','numba>=0.61.0']
 dynamic = ["version"]
 
+[project.optional-dependencies]
+notebooks = ["matplotlib", "jupyter"]
+dev = ["pytest",]
+
 [tool.hatch.version]
 source = "vcs"
 
