### 2025-11-07 a98d7e4
added explicitly the supported python versions to documentation
