### 2025-11-07 84c7aa2
added CONTRIBUTING.md
