+++ docs/help.rst
@@ -1,4 +1,0 @@
-Help / Bug reports / Contributing
-========================================
-
-For questions, bug reports, ideas for new features or contributions, please open an issue at `<https://github.com/gica3618/pythonradex/issues>`_.
