### 2025-11-07 gica3618/pythonradex#9 @GijsVermarien
Thank you for clearing it up, I forgot that Numba indeed takes some time to work with new releases.

I think the `pip install pythonradex[notebooks]` is useful indeed! Just to improve the user experience.
