### 2025-11-07 #9233 @gica3618
For the record, two issues opened by @GijsVermarien in the target repository where addressed:

- gica3618/pythonradex#9
- gica3618/pythonradex#10
