### 2025-11-07 #9233 @pscicluna
@gica3618 You've put together a great package and paper, here are some initial comments from me on the paper:

L12: I realise this is a little pedantic, but is it really fair to call it "pure python" when numpy and scipy are calling in to C/Fortran routines? I do get what you mean (you haven't explicitly called in to compiled code) however, so feel free to ignore me.
L50: It might be useful for non-expert readers to include a brief description of the advantages and limitations of the LVG approximation.
L60: Like @GijsVermarien I think it would be great to add some plots comparing the different codes.
L61-66:
 - I think it would be good to include more details on this benchmark. Are you using `multiprocessing` with `pythonradex` to fill all CPU schedules, in which case this would not be an entirely fair comparison, or is the performance improvement coming from better vectorisation with `numba`+`numpy` comapared to un-optimised Fortran? In the latter case, this is a real advantage for pythonradex, since it is much easier to get this performance uplift than it is by re-writing RADEX.
 - Would it be possible to plot how performance scales with number of CPU cores? This is assuming that `multiprocessing` is not being used.
 - I would also be curious to see how performance compares in Julia, although I can see why this might be beyond the scope of this work
L75-82: Again, I think a plot demonstrating the relative importance of this correction would be useful.

I'm still testing the code itself out and might be back with a few comments on that side too.
