### 2025-11-10 47c22b5
paper updates
