+++ joss_paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-A common task in astronomical research is to estimate the physical parameters (temperature, mass, density etc.) of a gas by using observed line emission. This often requires a calculation of how the radiation propagates via emission and absorption ("radiative transfer"). In radio and infrared astronomy, the Fortran code RADEX [@vanderTak2007] is a popular tool to solve the non-LTE radiative transfer of a uniform medium in a simplified geometry. I present a python implementation of RADEX: pythonradex. Written in pure python, it provides an easy and intuitive user interface as well as additional functionality not included in RADEX (continuum effects and overlapping lines).
+A common task in astronomical research is to estimate the physical parameters (temperature, mass, density etc.) of a gas by using observed line emission. This often requires a calculation of how the radiation propagates via emission and absorption ("radiative transfer"). In radio and infrared astronomy, the Fortran code `RADEX` [@vanderTak2007] is a popular tool to solve the non-LTE radiative transfer of a uniform medium in a simplified geometry. I present a python implementation of `RADEX`: `pythonradex`. Written in python, it provides an easy and intuitive user interface as well as additional functionality not included in `RADEX` (continuum effects and overlapping lines).
 
 # Statement of need
 
@@ -30,52 +30,58 @@ Modern facilities such as the Atacama Large Millimeter/submillimeter Array (ALMA
 
 To interpret such data, a radiative transfer calculation is typically used (see @Rybicki1985 for an introduction to radiative transfer). For a given set of input parameters describing the source (temperature, density, geometry, etc.), one calculates the amount of radiation reaching the telescope. The input parameters are then adjusted such that the predicted flux matches the observations.
 
-If the medium is dense enough, local thermodynamic equilibrium (LTE) maybe be assumed. This considerably simplifies the radiative transfer calculation. A non-LTE calculation is considerably more complex and computationally expensive because the fractional population of the molecular energy levels needs to be calculated explicitly.
+If the medium is dense enough, local thermodynamic equilibrium (LTE) maybe be assumed. This considerably simplifies the radiative transfer calculation. A non-LTE calculation is considerably more complex and computationally expensive **because the fractional population of the molecular energy levels needs to be solved for numerically [e.g. @vanderTak2007]. Typically, an iterative approach is used: from a first guess of the level populations, the radiation field is calculated. This radiation field is then used to solve for updated level populations. The iterations continue until convergence is reached.**
 
-Various codes are available to solve the radiative transfer. Codes solving the radiative transfer in 3D are used for detailed calculations of sources with well-known geometries. Examples include RADMC-3D [@Dullemond2012] and LIME [@Brinch2010]. However, a full 3D calculation is often too computationally expensive if a large parameter space needs to be explored, in particular in non-LTE. 1D codes that quickly provide an approximate solution are a commonly used alternative. In this respect, the 1D non-LTE code RADEX [@vanderTak2007] has gained considerable popularity: as of February 28, 2025, the RADEX paper by @vanderTak2007 has 1361 citations. The Fortan code RADEX solves the radiative transfer of a uniform medium using an escape probability formalism.
[…]
