+++ joss_paper/paper.bib
@@ -205,3 +205,42 @@ series = {LLVM '15}
  publisher     = {Springer Science and Business Media {LLC}},
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
+
+@Inbook{Elitzur1992,
+author="Elitzur, Moshe",
+title="Basic Background Concepts",
+bookTitle="Astronomical Masers",
+year="1992",
+publisher="Springer Netherlands",
+address="Dordrecht",
+pages="4--46",
+abstract="On a microscopic level a gas is comprised of material particles at various levels of internal excitations, moving randomly in space (executing thermal or kinetic motions). The problem of the kinetic theory of gases is to determine the distributions of the thermal velocities and of the level populations. This task is complicated by the fact that these distributions are determined by the combined effects of a number of competing mechanisms, including collisions between the particles as well as various radiative processes.",
+isbn="978-94-011-2394-5",
+doi="10.1007/978-94-011-2394-5_2",
+url="https://doi.org/10.1007/978-94-011-2394-5_2"
+}
+
+@ARTICLE{Schoier2005,
+       author = {{Sch{\"o}ier}, F.~L. and {van der Tak}, F.~F.~S. and {van Dishoeck}, E.~F. and {Black}, J.~H.},
+        title = "{An atomic and molecular database for analysis of submillimetre line observations}",
+      journal = {\aap},
+     keywords = {astronomical data bases: miscellaneous, atomic data, molecular data, radiative transfer, ISM: atoms, ISM: molecules, Astrophysics},
+         year = 2005,
+        month = mar,
+       volume = {432},
+       number = {1},
+        pages = {369-379},
+          doi = {10.1051/0004-6361:20041729},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0411110},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2005A&A...432..369S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@misc{EMAA,
+  title = {EMAA},
+  url = {https://emaa.osug.fr/},
+  year = {2021},
+  publisher = {UGA, CNRS, CNRS-INSU, OSUG}
+}
