+++ joss_paper/paper.md
@@ -62,11 +62,7 @@ The python programming language is now very widely used in astronomy. Still, no 
 
 ## Output flux
 
-<<<<<<< HEAD
-`RADEX` computes line fluxes based on a "background subtracted" intensity given by $(B_\nu(T_\mathrm{ex})-I_\mathrm{bg})(1-e^{-\tau_\nu})$, where $B_\nu$ is the Planck function, $T_\mathrm{ex}$ the excitation temperature, $I_\mathrm{bg}$ the external background and $\tau_\nu$ the optical depth. This may or may not be the right quantity to be compared to observations. In contrast, `pythonradex` does not apply any observational correction, thus giving more flexibility to the user.
-=======
 RADEX computes line fluxes based on a "background subtracted" intensity given by $(B_\nu(T_\mathrm{ex})-I_\mathrm{bg})(1-e^{-\tau_\nu})$, where $B_\nu$ is the Planck function, $T_\mathrm{ex}$ the excitation temperature, $I_\mathrm{bg}$ the external background and $\tau_\nu$ the optical depth. This may or may not be the right quantity to be compared to observations (for example, it is not appropriate when considering data from interferometers like ALMA). pythonradex does not apply any observational correction, giving the user the flexibility to decide how the computed fluxes are compared to observations.
->>>>>>> 5b36909e7b502d0bf5308d0d7167b1f5debd0cfa
 
 ## Flux for spherical geometry
 
