### 2025-11-18 5400606
paper updates and manual benchmark updates
