+++ joss_paper/paper.bib
@@ -242,5 +242,6 @@ archivePrefix = {arXiv},
   title = {EMAA},
   url = {https://emaa.osug.fr/},
   year = {2021},
-  publisher = {}
+  publisher = {UGA, CNRS, CNRS-INSU, OSUG},
+  doi = {10.17178/EMAA}
 }
