+++ joss_paper/code/molpop-cep-master/Source/molpop.f90
@@ -1,0 +1,269 @@
+program molpop
+use io_molpop
+use sol_molpop
+use global_molpop
+use maths_molpop
+use solution_cep, only: solve_with_cep
+implicit none
+!
+!  ---------------   MOLECULAR LEVEL POPULATIONS SOLVER   ---------------
+!
+!  Main subroutines:
+!
+!    const  - define mathematical and physical constants
+!    input - read input data
+!    data  - load molecular data
+!    first  - get first solution
+!    solve  - scale the dimensions and solve the level equations
+!
+!  Main variables:
+!
+!        n - number of levels
+!    kbeta - the escape probability used:
+!            = 0 - Sobolev;
+!             -1 - LVG slab, from Scoville & Solomon 1974, ApJ, 187, L67
+!              1 - sphere, from van der Tak+ 2007, A&A 468, 627, eq. 19
+!              2 - slab, version #2 KROLIK AND MCKEE, AP. J. SUPP. 37, P459 (1978)
+!                        (slab now uses line center tau instead of mean tau)
+!    kbeta = 3     Full CEP calcualtion
+!        r - dimension(cm):  radius in Sobolev approximation
+!            and thickness for slab
+!        v - velocity in Sobolev approximation, Doppler linewidth otherwise;
+!            entered in km/sec and then scaled to cm/sec
+!      sat - saturation parameter
+!            = 1 - maser affects level populations
+!              0 - no maser effects on level populations
+!     nmax - max # of iterations allowed
+!        T - kinetic temperature
+!     nh2  - h2 density (cm**-3)
+!     nmol - molecule density (cm**-3)
+!     mcol - molecular column density (cm-2/kms)
+!    tcool - total cooling rate
+! mol_mass - the molecular mass
+!     kprt - printing index for output
+!    kcool - printing empahsizes masing (0) or cooling (1)
+!   ksolpr - printing index for solve
+!   newtpr - printing index for newton
+!    inprt - printing index for input
+!   kfirst - printing index for initial guess
+!    nbig  - the number of cooling lines that will be printed
+!
+!
+!     molecular data
+!
+!     g     - statistical weights;
+!     a     - Einstein coefficients;
+!     c     - collision rates;
+!     we    - weight factors
+!
+!     fr    -  state energy (in cm**-1)
+!     ti    -  state energy (in K)
+!     freq  -  transition frequencies in Hz
+!     wl    -  wavelength in microns
+!     tij   -  transition frequencies in K
+!     boltz -  Boltzmann factor for state
+!     gap   -  Boltzmann factor for transition
+!     taux  -  multipliers for optical depths: tau = r*taux*delta(x)
+!     ems   -  multipliers for line emissivities
+!     rad   -  external radiation intensity in the lines
+!
+!        x - population per sublevel divided by Boltzmann factor
+!           (analogous to departure coefficient);
+!
+!    we(i) = g(i)*exp(-e(i)/kT);
+!     n(i) = nmol*x(i)*we(i) is the population of level i;
+!   pop(i) = x(i)*exp(-e(i)/kT) is the (relative) population per
+!            sub-level
+!
+!            a and c are scaled by we, so that
+!                 n(i) a(i,j) = nmol x(i) (we(i) a(i,j))
+!                 and similarly for n(i)*c(i,j)
+!
+!  The equations to solve are F(x1,....xn) = 0;
+!
+!        D - the jacobian matrix of the equations.
+!
+
+
+    logical quit, stat
+    character*168 fn_inp, fn_out
+    integer io1,I, Tcount, iunit, NHcount
+    real(kind=8), allocatable :: x(:),f(:),d(:,:),cool(:,:)
+    real(kind=8) :: NHarray
[…]

[…]
