### 2025-11-18 faf1dbb
paper updates
