### 2025-11-18 a770acd
paper updates
