### 2025-11-18 c56d442
paper updates
