+++ joss_paper/code/compare_pythonradex_RADEX.py
@@ -28,13 +28,13 @@ ext_background = helpers.generate_CMB_background()
 geometry = "LVG slab"
 geometry_radex = "LVG slab"
 # geometry = "uniform sphere"
-# geometry_radex = "uniform sphere"
+# geometry_radex = "static sphere"
 
 #don't use LVG sphere because RADEX uses a different formula for the escape
 #probability
-assert geometry == geometry_radex != "LVG sphere"
+assert geometry != "LVG sphere" and geometry_radex != "LVG sphere"
 #also RADEX does not support uniform slab
-assert geometry == geometry_radex != "uniform slab"
+assert geometry != "uniform slab"
 
 cloud = radiative_transfer.Cloud(datafilepath=datafilepath,geometry=geometry,
                                  line_profile_type=line_profile_type,
@@ -99,10 +99,10 @@ solid_angle = 1
 
 grid_size = 25
 N_grid = np.logspace(13,18,grid_size)*constants.centi**-2
-Tkin_grid = np.logspace(np.log10(5),np.log10(1000),grid_size)
+Tkin_grid = np.logspace(np.log10(5),np.log10(500),grid_size)
 N_GRID,TKIN_GRID = np.meshgrid(N_grid,Tkin_grid,indexing='ij')
 
-nH2_grid = np.array((1e3,1e4,1e6,1e8))*constants.centi**-3
+nH2_grid = np.array((1e2,2e3,5e4,1e6))*constants.centi**-3
 flux_residuals = []
 
 v = np.linspace(-width_v*3,width_v*3,30)
@@ -120,20 +120,20 @@ for i,nH2 in enumerate(nH2_grid):
     coll_densities = {'ortho-H2':nH2/2,'para-H2':nH2/2}
     for j,N in enumerate(N_grid):
         for k,Tkin in enumerate(Tkin_grid):
+            print(f"nH2 = {nH2/constants.centi**-3:.2g} cm-3,"
+                  +f" N = {N/constants.centi**-2:.2g}, Tkin={Tkin:.2g} K")
             radex_model = compute_RADEX_model(
                               N=N,Tkin=Tkin,coll_partner_densities=coll_densities,
                               transitions=[plot_trans,])
+            radex_model = radex_model[0][0],radex_model[1][0]
             pythonradex_model = compute_pythonradex_model(
-                                     N=N,Tkin=Tkin,coll_partner_densities=coll_densities)
+                                     N=N,Tkin=Tkin,
+                                     coll_partner_densities=coll_densities)
+            pythonradex_model = pythonradex_model[0][plot_trans_index],\
+                                pythonradex_model[1][plot_trans_index]
             models = {"radex":radex_model,"pythonradex":pythonradex_model}
             for ID,model in models.items():
                 Tex,tau_nu0 = model
-                if ID == "pythonradex":
-                    Tex,tau_nu0 = Tex[plot_trans_index],tau_nu0[plot_trans_index]
-                elif ID == "radex":
-                    Tex,tau_nu0 = Tex[0],tau_nu0[0]
-                else:
-                    raise RuntimeError
                 S = helpers.B_nu(nu=nu,T=Tex)
                 phi_nu = plot_trans.line_profile.phi_nu(nu)
                 tau_nu = phi_nu/np.max(phi_nu)*tau_nu0
@@ -153,8 +153,10 @@ for i,nH2 in enumerate(nH2_grid):
     ax.set_yscale("log")
     ax.set_xlim(np.min(N_grid)/constants.centi**-2,np.max(N_GRID)/constants.centi**-2)
[…]
