### 2025-11-18 dd45878
paper updates
