### 2025-11-18 #9233 @gica3618
Dear @GijsVermarien, @pscicluna, @andresmegias, thank you for reviewing the paper. I will address your comments in order, beginning with @GijsVermarien's comments.

> L8 - You mention we can derive physical parameters, I understand that we can derive temperature and density using a radiative transfer code, however mass is not that easy to derive without many more assumptions?

I do not see a fundamental difference between deriving temperature, density or mass / column density. All are input parameters for the radiative transfer calculation that can be fitted wrt to some data. Of course, if the emission is optically thick, one can generally only derive a lower limit for the mass / column density, since increasing the mass more and more does (almost) not change the emission.

> L28 - A line explaining what this iterative solving process looks like for non-LTE systems would be useful.
L50 - Here it would be useful to briefly explain the LVG assumption.
Figure 1 - Could you add the black dotted lines to the legend?

All done.

> L60 - I would like to see a figure that illustrates this comparison

- In the figure showing the calculation of overlapping lines, I added a comparison to `MOLPOP-CEP`. Good agreement is seen. Interestingly, if the treatment of overlapping lines is turned off, the two codes somewhat disagree.
- I also added a new figure that compares `pythonradex` to `RADEX` using CO 3-2 as an example, illustrating the good agreement.

> General remark: you could add the citation for the collisional data files from LAMBDA([10.1051/0004-6361:20041729](https://ui.adsabs.harvard.edu/link_gateway/2005A&A...432..369S/doi:10.1051/0004-6361:20041729)) and EMAA(http://dx.doi.org/10.17178/EMAA) database since they are a crucial part of your workflow.

Done.
