### 2025-11-18 e083936
paper updates
