### 2025-11-19 c406fa5
work on performance tests
