+++ manual_tests_and_benchmarks/performance_tests/pythonradex_vs_radex/check_radex_CPU_bash.sh
@@ -1,0 +1,5 @@
+#in terminal, do something like /usr/bin/time -v ./check_radex_CPU_bash.sh
+
+for i in $(seq 1 1000); do
+    ../../../tests/Radex/bin/radex_static_sphere < radex_test_preformance.inp > /dev/null
+done
