+++ manual_tests_and_benchmarks/general.py
@@ -8,7 +8,9 @@ Created on Mon Feb 10 20:07:06 2025
 
 import os
 
-lamda_data_folder = '/home/gianni/science/projects/code/pythonradex/tests/LAMDA_files'
+this_filepath = os.path.abspath(__file__)
+dirpath = os.path.dirname(this_filepath)
+lamda_data_folder = os.path.join(dirpath,'../tests/LAMDA_files')
 
 def datafilepath(filename):
     return os.path.join(lamda_data_folder,filename)
