### 2025-11-19 ad7a68a
added more time information
