+++ manual_tests_and_benchmarks/performance_tests/pythonradex_vs_radex/check_radex_CPU_bash.sh
@@ -1,3 +1,4 @@
+#this is just to check how much CPU radex uses
 #in terminal, do something like /usr/bin/time -v ./check_radex_CPU_bash.sh
 
 for i in $(seq 1 1000); do
