+++ manual_tests_and_benchmarks/performance_tests/pythonradex_vs_radex/performance_comparison.py
@@ -29,12 +29,12 @@ radex_executable = '../../../tests/Radex/bin/radex_static_sphere'
 # geometry = 'LVG slab'
 # radex_executable = '../../../tests/Radex/bin/radex_LVG_slab'
 
-n_grid_elements = 15
+n_grid_elements = 20
 
-data_filename = 'co.dat'
-colliders = ['para-H2','ortho-H2']
-log_N_limits = 13+4,18+4
-Tmin,Tmax = 20,250
+# data_filename = 'co.dat'
+# colliders = ['para-H2','ortho-H2']
+# log_N_limits = 13+4,18+4
+# Tmin,Tmax = 20,250
 #narrow range where RADEX does not throw warnings:
 # log_N_limits = 16+4,16.3+4
 # Tmin,Tmax = 50,51
