### 2025-11-20 bf6f304
updates for performance tests
