### 2025-11-20 a9e910d
updates for performance tests
