+++ manual_tests_and_benchmarks/performance_tests/pythonradex_vs_radex/performance_comparison.py
@@ -7,6 +7,9 @@ Created on Tue Nov 18 14:50:22 2025
 """
 
 #this script is essentially a simplified version of radex_vs_pythonradex_performance.py
+#IMPORTANT: if you want to delete the numba compilation (__pycache__), you need to
+#run this in an environment where pythonradex is installed in editable mode, e.g.
+#the virtual environment virt_env_for_testing
 
 import sys
 sys.path.append('../..')
@@ -18,7 +21,7 @@ import numpy as np
 import time
 import itertools
 import mini_radex_wrapper
-
+import shutil
 
 ext_background = helpers.generate_CMB_background(z=0)
 
@@ -28,101 +31,90 @@ ext_background = helpers.generate_CMB_background(z=0)
 geometry = 'uniform sphere'
 # geometry = 'LVG slab'
 
-n_grid_elements = 15
+n_grid_elements = 20
 
-# data_filename = 'co.dat'
-# colliders = ['para-H2','ortho-H2']
-# log_N_limits = 13+4,18+4
-# Tmin,Tmax = 20,250
-#narrow range where RADEX does not throw warnings:
-# log_N_limits = 16+4,16.3+4
-# Tmin,Tmax = 50,51
-
-# data_filename = 'hco+.dat'
-# colliders = ['H2',]
-# log_N_limits = 10+4,14+4
-# Tmin,Tmax = 20,250
-
-# data_filename = 'so@lique.dat'
-# colliders = ['H2',]
-# log_N_limits = 10+4,12+4
-# Tmin,Tmax = 60,250
-
-data_filename = 'c.dat'
-colliders = ['para-H2','ortho-H2']
-log_N_limits = 12+4,18+4
-Tmin,Tmax = 60,250
-
-
-# #ATTENTION: if no H2 is given, RADEX just puts 1e5 cm-3 by default! WTF!
-#RADEX does other strange things with the H2 density:
-#see line 112 in io.f, line 168 in io.f, and line 225 in readdata.f
-for collider in colliders:
-    assert 'H2' in collider
+grids = {"CO":{'data_filename':'co.dat','colliders':['para-H2','ortho-H2'],
+               "log_N_limits" :(13+4,18+4),"Tmin":20,'Tmax':250},
+         #narrow range where RADEX does not throw warnings:
+         "CO narrow range":{'data_filename':'co.dat','colliders':['para-H2','ortho-H2'],
+                            "log_N_limits" :(16+4,16.3+4),"Tmin":50,'Tmax':51},
+         "HCO+":{'data_filename':'hco+.dat','colliders':['H2',],
+                 "log_N_limits" :(10+4,14+4),"Tmin":20,'Tmax':250},
+         "SO":{'data_filename':'so@lique.dat','colliders':['H2',],
+               "log_N_limits" :(10+4,12+4),"Tmin":60,'Tmax':250},
+         "C":{'data_filename':'c.dat','colliders':['para-H2','ortho-H2'],
+              "log_N_limits" :(12+4,18+4),"Tmin":60,'Tmax':250}
+         }
 
 #actually, RADEX assumes rectangular, but then converts it Gaussian for the line flux
 line_profile_type = 'rectangular'
 width_v = 1*constants.kilo
 
-datafilepath = os.path.join(general.lamda_data_folder,data_filename)
-
 radex_input_file = 'radex_test_preformance.inp'
 radex_output_file = 'radex_test_preformance.out'
 
-N_values = np.logspace(*log_N_limits,n_grid_elements)
-coll_density_values = np.logspace(3,5,n_grid_elements)/constants.centi**3
-Tkin_values = np.linspace(Tmin,Tmax,n_grid_elements)
+def remove_pythonradex_cache():
[…]
