### 2025-11-21 05d0b4d
figure for joss paper
