### 2025-11-21 c01e1d6
paper updates
