+++ joss_paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'pythonradex: a python implementation of the non-LTE radiative transfer code RADEX with additional functionality'
+title: 'pythonradex: a fast Python re-implementation of RADEX with extended functionality'
 tags:
   - python
   - astronomy
@@ -26,7 +26,7 @@ A common task in astronomical research is to estimate the physical parameters (t
 
 # Statement of need
 
-Modern facilities such as the Atacama Large Millimeter/submillimeter Array (ALMA) or the James Webb Space Telescope (JWST) are providing a wealth of line emission data at radio and infrared wavelengths. These data are crucial to constrain the physical and chemical properties of various astrophysical environments.
+Modern **astronomical** facilities such as the Atacama Large Millimeter/submillimeter Array (ALMA) or the James Webb Space Telescope (JWST) are providing a wealth of line emission data at radio and infrared wavelengths. These data are crucial to constrain the physical and chemical properties of various astrophysical environments.
 
 To interpret such data, a radiative transfer calculation is typically used (see @Rybicki1985 for an introduction to radiative transfer). For a given set of input parameters describing the source (temperature, density, geometry, etc.), one calculates the amount of radiation reaching the telescope. The input parameters are then adjusted such that the predicted flux matches the observations.
 
@@ -55,9 +55,9 @@ The python programming language is now very widely used in astronomy. Still, no 
 ![Spectrum of HCN around 177.3 GHz computed with `pythonradex` **and `MOLPOP-CEP` for a static slab. Good agreement is found when treating line overlap. Interestingly, the spectra differ somewhat when ignoring overlap.** The positions and widths of the individual hyperfine components are illustrated by the black dotted lines.\label{fig:HCN_spectrum}](code/HCN_spec.pdf)
 
 
-# Performance
+# Performance advantage
 
-`pythonradex` is optimised for the use case of a parameter space exploration. On a laptop with four i7-7700HQ cores (eight virtual cores), running a grid of CO models was three times faster with `pythonradex` compared to `RADEX`. However, the CPU usage of `pythonradex` was much higher than `RADEX` during this test. Still, it demonstrates that `pythonradex` might be faster than `RADEX` depending on the setup.
+**Both `pythonradex` and `RADEX` are single-threaded. To compare their performance, we consider the calculation of a grid of models over a parameter space spanning 20 values in each of kinetic temperature, column density and H$_2$ density (i.e. a total of 8000 models). We consider a few different molecules: C (small number of levels and transitions), SO (large number of levels and transitions) as well as CO and HCO$^+$ (intermediate). On a laptop with i7-7700HQ cores running on Ubuntu 22.04, `pythonradex` computed the model grid faster than `RADEX` by factors 2 (C), 6 (SO), 6 (CO) and 3 (HCO$^+$). Running the same test on the Multi-wavelength Data Analysis System (MDAS) operated by NAOJ (Rocky Linux 8.9 with AMD EPYC 7543 CPUs) resulted in a even larger performance advantage: `pythonradex` calculated faster by factors of 18 (C), 14 (SO), 12 (CO) and 14 (HCO$^+$).**
 
 # Additional differences between RADEX and pythonradex
 
@@ -67,11 +67,14 @@ RADEX computes line fluxes based on a "background subtracted" intensity given by
 
 ## Flux for spherical geometry
 
[…]
