### 2025-11-21 3361d43
reverted changes to sphere flux
