+++ joss_paper/paper.md
@@ -57,7 +57,7 @@ The python programming language is now very widely used in astronomy. Still, no 
 
 # Performance advantage
 
-**Both `pythonradex` and `RADEX` are single-threaded. To compare their performance, we consider the calculation of a grid of models over a parameter space spanning 20 values in each of kinetic temperature, column density and H$_2$ density (i.e. a total of 8000 models). We consider a few different molecules: C (small number of levels and transitions), SO (large number of levels and transitions) as well as CO and HCO$^+$ (intermediate). On a laptop with i7-7700HQ cores running on Ubuntu 22.04, `pythonradex` computed the model grid faster than `RADEX` by factors 2 (C), 6 (SO), 6 (CO) and 3 (HCO$^+$). Running the same test on the Multi-wavelength Data Analysis System (MDAS) operated by NAOJ (Rocky Linux 8.9 with AMD EPYC 7543 CPUs) resulted in a even larger performance advantage: `pythonradex` calculated faster by factors of 18 (C), 14 (SO), 12 (CO) and 14 (HCO$^+$).**
+**Both `pythonradex` and `RADEX` are single-threaded. To compare their performance, we consider the calculation of a grid of models over a parameter space spanning 20 values in each of kinetic temperature, column density and H$_2$ density (i.e. a total of 8000 models). We consider a few different molecules: C (small number of levels and transitions), SO (large number of levels and transitions) as well as CO and HCO$^+$ (intermediate). On a laptop with i7-7700HQ cores running on Ubuntu 22.04, `pythonradex` computed the model grid faster than `RADEX` by factors 2 (C), 6 (SO), 6 (CO) and 3 (HCO$^+$). Running the same test on the Multi-wavelength Data Analysis System (MDAS) operated by the National Astronomical Observatory of Japan (Rocky Linux 8.9 with AMD EPYC 7543 CPUs) resulted in a even larger performance advantage: `pythonradex` calculated faster by factors of 18 (C), 14 (SO), 12 (CO) and 14 (HCO$^+$).**
 
 # Additional differences between RADEX and pythonradex
 
@@ -67,14 +67,11 @@ RADEX computes line fluxes based on a "background subtracted" intensity given by
 
 ## Flux for spherical geometry
 
-Regardless of the adopted geometry, `RADEX` always uses the flux formula for a slab geometry **($F_\nu=B_\nu(T_\mathrm{ex})(1-e^{-\tau_\nu})$)**, resulting in inconsistencies. Consider the optically thin limit where the total flux (in [W/m$^2$]) for a sphere is simply given by
+Regardless of the adopted geometry, `RADEX` always uses the flux formula for a slab geometry, resulting in inconsistencies. Consider the optically thin limit where the total flux (in [W/m$^2$]) for a sphere is simply given by
 \begin{equation}
 F_\mathrm{thin} = V_\mathrm{sphere}n_2A_{21}\Delta E \frac{1}{4\pi d^2}
 \end{equation}
[…]
