### 2025-11-21 cdae0a3
paper updates
