### 2025-11-26 a0f5195
updated LVG sphere flux: updated performance benchmarks
