+++ tests/test_escape_probability.py
@@ -37,3 +37,16 @@ def test_flux_uniform_sphere():
                           tau_nu=np.array((limit_tau_nu+epsilon_tau_nu,)),
                           source_function=source_function,solid_angle=solid_angle)
     assert np.isclose(flux_Taylor,flux_analytical,rtol=0.05,atol=0)
+
+def test_flux_LVG_sphere():
+    V = 1
+    v = np.linspace(-2*V,2*V,100)
+    nu0 = 100*constants.giga
+    nu = nu0*(1-v/constants.c)
+    flux = escape_probability.UniformLVGSphere.compute_flux_nu(
+               tau_nu=1,source_function=1,solid_angle=1,nu=nu,nu0=nu0,V=V)
+    zero_region = np.abs(v) > V
+    assert np.any(zero_region)
+    assert np.any(~zero_region)
+    assert np.all(flux[~zero_region]>0)
+    assert np.all(flux[zero_region]==0)
