+++ manual_tests_and_benchmarks/emerging_flux/thin_flux_sphere_vs_RADEX.py
@@ -7,10 +7,12 @@ Created on Tue May 28 17:10:58 2024
 """
 
 #RADEX uses I=S(1-exp(-tau)) even for a sphere, while I use the formula from
-#Osterbrock. The two formulae give different results particularly for the optically
-#thin case (see compare_emerging_flux_formulae.py)
+#Osterbrock (for static sphere), and another formular for LVG sphere.
+#The two formulae give different results
+#particularly for the optically thin case (see compare_emerging_flux_formulae.py)
 #here I test explicitly that the Osterbrock formula indeed gives the flux expected
 #in the optically thin case
+
 
 import numpy as np
 from scipy import constants
@@ -25,7 +27,9 @@ r = 1*constants.au
 d = 1*constants.parsec
 n = 20/constants.centi**3
 width_v = 1*constants.kilo
-line_profile_type = 'Gaussian'
+#actually RADEX does calculation using rectangular, but then applies correction
+#factor for Gaussian... but fortunately, for optically thin case, it doesn't matter
+line_profile_type = 'rectangular'
 Tex = 30
 trans_index = 1
 #make it LTE for the RADEX wrapper so that Tkin=Tex
@@ -54,30 +58,43 @@ volume = 4/3*r**3*np.pi
 solid_angle = r**2*np.pi/d**2
 #W/m2
 thin_flux = volume*n*level_pop[trans.up.number]*trans.A21*trans.Delta_E/(4*np.pi*d**2)
-uniform_sphere = escape_probability.UniformSphere()
+print(f"analytical flux: {thin_flux:.3g} W/m2")
 source_func = helpers.B_nu(nu=nu,T=Tex)
-flux_kwargs = {'tau_nu':tau_nu,'source_function':source_func,'solid_angle':solid_angle}
-flux_pythonradex = uniform_sphere.compute_flux_nu(**flux_kwargs)
-flux_pythonradex = np.trapezoid(flux_pythonradex,nu)
-flux_0D = escape_probability.Flux1D()
-flux_RADEX = flux_0D.compute_flux_nu(**flux_kwargs)
-flux_RADEX = np.trapezoid(flux_RADEX,nu)
 
-radex_input = radex_wrapper.RadexInput(
-                     data_filename=datafilepath,
-                     frequency_interval=frequency_interval,Tkin=Tex,
-                     coll_partner_densities=coll_partner_densities,
-                     T_background=0,column_density=n*2*r,
-                     Delta_v=width_v)
-radex_wrap = radex_wrapper.RadexWrapper(geometry='static sphere')
-radex_wrap.compute(radex_input)
-output = radex_wrap.compute(radex_input)
-#turns out to calculate the flux from the antenna temperature, RADEX simply
-#integrates the antenna temperature intensity over all solid angles (not sure why...)
-#so need to scale by the solid angle of the source
-#see line 334 in io.f
-radex_wrapper_flux = output['flux']*solid_angle/(4*np.pi)
+geometries = {"uniform sphere":escape_probability.UniformSphere(),
+              "lvg sphere":escape_probability.UniformLVGSphere(),
+              "uniform sphere RADEX":escape_probability.UniformSphereRADEX(),
+              "lvg sphere RADEX":escape_probability.LVGSphereRADEX()}
+flux_kwargs_template = {'tau_nu':tau_nu,'source_function':source_func,
+                        'solid_angle':solid_angle}
+flux_kwargs = {geo_name:flux_kwargs_template.copy() for geo_name in
+               geometries.keys()}
+flux_kwargs["lvg sphere"]["nu"] = nu
+flux_kwargs["lvg sphere"]["nu0"] = trans.nu0
+flux_kwargs["lvg sphere"]["V"] = width_v/2
 
[…]
