+++ manual_tests_and_benchmarks/emerging_flux/compare_emerging_flux_formula.py
@@ -6,55 +6,81 @@ Created on Tue May 28 10:58:53 2024
 @author: gianni
 """
 
-#RADEX uses B(Tex)*(1-exp(-tau)) for the emerging flux even for a uniform sphere
-#but looking at Osterbrock, this is not correct, but let's compare how big the difference
-#actually is
+#RADEX uses B(Tex)*(1-exp(-tau)) for the emerging flux even for a static or LVG sphere
+#this is not correct, but let's check how big the difference actually is
 
+#assume rectangular line profile, to include LVG sphere
 
 import numpy as np
 import matplotlib.pyplot as plt
+from scipy import constants
+from pythonradex import helpers
 
-def flux_uniform_sphere(tau_nu,source_function,solid_angle):
-    return 2*np.pi*source_function/tau_nu**2\
-               *(tau_nu**2/2-1+(tau_nu+1)*np.exp(-tau_nu)) *solid_angle/np.pi
+nu0 = 100*constants.giga
+Tex = 30
+source_function_nu0 = helpers.B_nu(nu=nu0,T=Tex)
+solid_angle = (1*constants.au/(1*constants.parsec))**2
+width_v = 1*constants.kilo
 
-def flux_0D(tau_nu,source_function,solid_angle):
-    return source_function*(1-np.exp(-tau_nu))*solid_angle
+
+width_nu = width_v/constants.c*nu0
+V = width_v/2
+
+def flux_static_sphere(tau_nu0):
+    flux_nu0 = 2*np.pi*source_function_nu0/tau_nu0**2\
+               *(tau_nu0**2/2-1+(tau_nu0+1)*np.exp(-tau_nu0)) *solid_angle/np.pi
+    #assume rectangular profile:
+    return flux_nu0*width_nu
+
+def flux_lvg_sphere(tau_nu0):
+    #see LVG_sphere.pdf
+    return nu0/constants.c*solid_angle*source_function_nu0*(1-np.exp(-tau_nu0))\
+           *4/3*V
+
+def flux_0D(tau_nu0):
+    return source_function_nu0*(1-np.exp(-tau_nu0))*solid_angle*width_nu
 
 
 tau_values = np.logspace(-2,2,100)
-source_function = 1
-solid_angle = 1
 
-f_sphere = flux_uniform_sphere(tau_nu=tau_values,source_function=source_function,
-                               solid_angle=solid_angle)
-f_0D = flux_0D(tau_nu=tau_values,source_function=source_function,solid_angle=solid_angle)
-relative_diff = np.abs((f_sphere-f_0D)/f_sphere)
-ratio = f_sphere/f_0D
-#in the optically thin case, the ratio should be (volume of sphere) / (volume of cylinder)
+pythonradex_fluxes = {"static sphere":flux_static_sphere(tau_nu0=tau_values),
+                      "LVG sphere":flux_lvg_sphere(tau_nu0=tau_values)}
+f_0D = flux_0D(tau_nu0=tau_values)
+
+
+#in the optically thin case, the ratio should be (volume of cylinder) / (volume of sphere)
 #(because the 0D formula corresponds to intensity independent over the emitting area),
-#which is (4/3 r**3 pi) / (r**2*pi * 2r) = 2/3
-expected_ratio_thin = 2/3
+#which is  (r**2*pi * 2r) / (4/3 r**3 pi) = 3/2
+expected_ratio_thin = 3/2
 
 
-fig,axes = plt.subplots(3)
-axes[0].plot(tau_values,f_sphere,label='sphere')
-axes[0].plot(tau_values,f_0D,label='0D (RADEX)')
[…]
