### 2025-11-28 4a09eae
paper updates
