### 2025-11-28 16170c3
paper updates
