### 2025-11-28 3fd846f
paper updates
