### 2025-11-28 #9233 @gica3618
Hi @pscicluna, thanks for your comments!

> L12: I realise this is a little pedantic, but is it really fair to call it "pure python" when numpy and scipy are calling in to C/Fortran routines? I do get what you mean (you haven't explicitly called in to compiled code) however, so feel free to ignore me.

This is a fair point. What I wanted to say with "pure python" is that this is not a wrapper of `RADEX`. But I guess that's clear anyway. I removed "pure".

>L50: It might be useful for non-expert readers to include a brief description of the advantages and limitations of the LVG approximation.

I added a brief description of the LVG assumption, but I think a discussion of the advantages and limitations might be too much, given that the paper already exceeds the word count limit of JOSS.

> L60: Like @GijsVermarien I think it would be great to add some plots comparing the different codes.

I added plots and text.

> L61-66: I think it would be good to include more details on this benchmark. Are you using multiprocessing with pythonradex to fill all CPU schedules, in which case this would not be an entirely fair comparison, or is the performance improvement coming from better vectorisation with numba+numpy comapared to un-optimised Fortran? In the latter case, this is a real advantage for pythonradex, since it is much easier to get this performance uplift than it is by re-writing RADEX. Would it be possible to plot how performance scales with number of CPU cores? This is assuming that multiprocessing is not being used. I would also be curious to see how performance compares in Julia, although I can see why this might be beyond the scope of this work

First, both `RADEX` and `pythonradex` are single-threaded. I confirmed this by running them with `GNU Time`. The CPU usage is 100% or less. Second, no multiprocessing is used in `pythonradex`.

I now ran a slightly more detailed test suite. I consider a parameter space grid with 20 elements in each of kinetic temperature, column density and H2 density (i.e. a total of 8000 models). I compared the time it takes for `RADEX` and `pythonradex` to compute the whole model grid, considering a few different molecules. I ran the test both on my laptop and on [MDAS](https://www.adc.nao.ac.jp/MDAS/mdas_e.html). I turns out that `pythonradex` always computes the grid faster than RADEX by up to a factor ~13. Note that for each model, I need to write a new input file for `RADEX` (this is done in python). However, I verified that the time it takes to write the input file is negligible compared to the execution of `RADEX`.

As for the comparison to Julia, there is [`Jadex`](https://github.com/autocorr/Jadex.jl) which claims a performance advantage of a factor ~110 over `RADEX`. I did not verify it, but I put a footnote in the paper.

I have re-written the section about the performance tests.

I am now quite convinced that pythonradex has a real performance advantage, so I would like to change the title of the paper to "pythonradex: a fast Python re-implementation of RADEX with extended functionality". @warrickball, besides changing the title in the pdf, do I need to change it anywhere else?

> L75-82: Again, I think a plot demonstrating the relative importance of this correction would be useful.

I added a plot. I have also rewritten that section.

Please have a look at the updated paper. All new text is in **bold**.
