### 2025-11-28 #9233 @GijsVermarien
Thank you for your updates, I read the latest version and think your added speed comparison and verification with MOLPOP-CEP round out the paper nicely and grant you the title "fast".

Regarding the comparison with radex in figure 1, is `pythonradex` consistently over or underpredicting w.r.t. `RADEX`?

Besides this, I think the paper and software is ready for publication when the other reviewers are happy as well.
