### 2025-12-01 a843ca4
paper updated
