### 2025-12-01 75775f8
paper updated
