### 2025-12-01 2f6ba7f
paper update
