### 2025-12-01 #9233 @gica3618
Hi @GijsVermarien,
>Regarding the comparison with radex in figure 1, is pythonradex consistently over or underpredicting w.r.t. `RADEX`?

That's a good point. It turns out that `pythonradex` is overpredicting wrt `RADEX`. I changed the figure so that this becomes clear. Please have a look.
