### 2025-12-01 8b8366b
changed uniform to static
