### 2025-12-02 905a3da
paper updates
