+++ joss_paper/code/compare_pythonradex_RADEX.py
@@ -16,25 +16,46 @@ import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib.colors import SymLogNorm
 
+#TODO either plot Tex instead of flux, or calculate the flux by using the level
+#populations, not tau (since tau given by RADEX contains a "correction" for
+#Gaussian line shape
+#TODO: check the values in the region where there is strong disagreement
 
-line_profile_type = "rectangular"
 data_filename = "co.dat"
-width_v = 1*constants.kilo
-datafilepath = os.path.join("../../tests/LAMDA_files",data_filename)
-
-T_background = 2.73
-ext_background = helpers.generate_CMB_background()
 #can only use LVG slab and uniform sphere
 geometry = "LVG slab"
 geometry_radex = "LVG slab"
 # geometry = "uniform sphere"
 # geometry_radex = "static sphere"
 
+#grid flux comparison
+plot_trans_index = 1
+grid_size = 25
+#grid_size = 3
+
+vmin = {"Tex":-10,"flux":-10}
+vmax = {"Tex":10,"flux":10}
+cmap = plt.get_cmap("PuOr").copy()
+cmap.set_bad("red")
+
+
+##############################################################
+
+#for consistency with RADEX:
+T_background = 2.73
+ext_background = helpers.generate_CMB_background()
+line_profile_type = "rectangular"
+
+solid_angle = 1
+width_v = 1*constants.kilo
+datafilepath = os.path.join("../../tests/LAMDA_files",data_filename)
+
 #don't use LVG sphere because RADEX uses a different formula for the escape
 #probability
 assert geometry != "LVG sphere" and geometry_radex != "LVG sphere"
 #also RADEX does not support uniform slab
 assert geometry != "uniform slab"
+
 
 cloud = radiative_transfer.Cloud(datafilepath=datafilepath,geometry=geometry,
                                  line_profile_type=line_profile_type,
@@ -44,11 +65,16 @@ def compute_pythonradex_model(N,Tkin,coll_partner_densities):
                             collider_densities=coll_partner_densities,
                             T_dust=0,tau_dust=0)
     cloud.solve_radiative_transfer()
-    return cloud.Tex,cloud.tau_nu0_individual_transitions
+    pop_up = [cloud.level_pop[t.up.number] for t in
+              cloud.emitting_molecule.rad_transitions]
+    pop_low = [cloud.level_pop[t.low.number] for t in
+               cloud.emitting_molecule.rad_transitions]
+    return cloud.Tex,pop_up,pop_low
 
 def compute_RADEX_model(N,Tkin,coll_partner_densities,transitions):
     Tex = []
-    tau_nu0 = []
+    pop_up = []
+    pop_low = []
     epsilon = 1*constants.giga
     for trans in transitions:
         freq_interval = radex_wrapper.Interval(min=trans.nu0-epsilon,
@@ -60,13 +86,14 @@ def compute_RADEX_model(N,Tkin,coll_partner_densities,transitions):
         wrapper = radex_wrapper.RadexWrapper(geometry=geometry_radex)
         results = wrapper.compute(radex_input)
         Tex.append(results["Tex"])
-        tau_nu0.append(results["tau"])
-    return Tex,tau_nu0
+        pop_up.append(results["pop_up"])
+        pop_low.append(results["pop_low"])
+    return Tex,pop_up,pop_low
 
 #plot comparison for a single model:
[…]
