### 2025-12-02 68ae81a
paper updates
