### 2025-12-02 2f6527c
paper updates
