### 2025-12-02 eeb9e20
paper updates
