### 2025-12-02 #9233 @gica3618
Hi @GijsVermarien, all,
Regarding the comparison between `pythonradex` and `RADEX`, there is a subtly regarding the fluxes. `RADEX` applies a "correction factor" to convert the rectangular line profile into a Gaussian when calculating optical depth and flux. So flux should not be compared directly between the two codes. Therefore, I updated the figure to compare excitation temperature (see updated paper). The alternative is to take the level populations predicted by the two codes and convert them to flux in a consistent way to allow for a fair comparison. I also did that, see attached figure that compares those fluxes. If you prefer the flux figure, please let me know.

[flux_pythonradex_vs_radex.pdf](https://github.com/user-attachments/files/23875301/flux_pythonradex_vs_radex.pdf)
