### 2025-12-02 #9233 @GijsVermarien
> Hi [@GijsVermarien](https://github.com/GijsVermarien), all, Regarding the comparison between `pythonradex` and `RADEX`, there is a subtly regarding the fluxes. `RADEX` applies a "correction factor" to convert the rectangular line profile into a Gaussian when calculating optical depth and flux. So flux should not be compared directly between the two codes. Therefore, I updated the figure to compare excitation temperature (see updated paper). The alternative is to take the level populations predicted by the two codes and convert them to flux in a consistent way to allow for a fair comparison. I also did that, see attached figure that compares those fluxes. If you prefer the flux figure, please let me know.
>
> [flux_pythonradex_vs_radex.pdf](https://github.com/user-attachments/files/23875301/flux_pythonradex_vs_radex.pdf)

Thank you for this plot, but I think its not very insightful to plot the relative error w.r.t pythonradex with this colormap. Might I suggest you keep this colormap, but instead plot the log distance between the two? So `log10(F_python_radex)-log10(F_radex)`. This is symmetric and suits your colormap much better with the middle being no deviation, you could even drop the red values. But this improved plot and analysis is a great addition, thank you!
