### 2025-12-03 #9233 @gica3618
Hi @GijsVermarien, sorry I could not catch the reason why you think the relative error wrt to `pythonradex` doesn't work well with the colormap. The relative error is also symmetric, and the middle (0%) means no deviation. The log ratio you suggest would also work, but I find it difficult to interpret, since one has to mentally take the value shown in the figure to the power of 10 to get a sense of what it means. Maybe simplest is just to show the ratio? Please see the attached figure showing simply F_pythonradex/F_radex (i.e. same as you propose, but not in log). As for the red points, some clipping is needed, otherwise the figure is completely dominated by a few pixels with very high deviations.

Also, do I understand correctly that you prefer a comparison of fluxes, rather than excitation temperatures? Fluxes are indeed what most users will be interested in, but as explained above, `pythonradex` and `RADEX` calculate fluxes in a different way. So in the paper, I would need to explain that the fluxes shown in the figure are not the fluxes computed by `RADEX`. Instead, the level populations computed by the two codes are converted to fluxes, and then compared.

[flux_pythonradex_vs_radex.pdf](https://github.com/user-attachments/files/23900985/flux_pythonradex_vs_radex.pdf)
[Tex_pythonradex_vs_radex.pdf](https://github.com/user-attachments/files/23900984/Tex_pythonradex_vs_radex.pdf)
