### 2025-12-04 ccaab13
radex wrapper for notebooks
