+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = ['scipy>=1.14','numpy>=2.0','numba>=0.61.0']
 dynamic = ["version"]
 
 [project.optional-dependencies]
-notebooks = ["matplotlib", "jupyter"]
+notebooks = ["matplotlib", "jupyter", "pandas"]
 dev = ["pytest",]
 
 [tool.hatch.version]
