### 2025-12-05 e328041
benchmark notebooks
