+++ manual_tests_and_benchmarks/performance_tests/flux_performance.py
@@ -16,7 +16,7 @@ import time
 #check how much time the flux calculation takes compared to the time it takes
 #to solve the radiative transfer
 
-geometry = 'uniform sphere'
+geometry = 'static sphere'
 line_profile_type = 'rectangular'
 width_v = 1*constants.kilo
 use_Ng_acceleration = True
