### 2025-12-08 6de0d2d
paper updates
