+++ joss_paper/code/compare_pythonradex_RADEX.py
@@ -14,12 +14,7 @@ from pythonradex import radiative_transfer,helpers
 import os
 import matplotlib.pyplot as plt
 import numpy as np
-from matplotlib.colors import SymLogNorm
 
-#TODO either plot Tex instead of flux, or calculate the flux by using the level
-#populations, not tau (since tau given by RADEX contains a "correction" for
-#Gaussian line shape
-#TODO: check the values in the region where there is strong disagreement
 
 data_filename = "co.dat"
 #can only use LVG slab and uniform sphere
@@ -28,13 +23,12 @@ geometry_radex = "LVG slab"
 # geometry = "uniform sphere"
 # geometry_radex = "static sphere"
 
-#grid flux comparison
+#for grid comparison:
 plot_trans_index = 1
 grid_size = 25
-#grid_size = 3
+#grid_size = 5
 
-vmin,vmax = -100,100
-linthresh = 1
+vmin,vmax = 0.95,1.05
 cmap = plt.get_cmap("PuOr").copy()
 cmap.set_bad("red")
 
@@ -53,7 +47,7 @@ datafilepath = os.path.join("../../tests/LAMDA_files",data_filename)
 #don't use LVG sphere because RADEX uses a different formula for the escape
 #probability
 assert geometry != "LVG sphere" and geometry_radex != "LVG sphere"
-#also RADEX does not support uniform slab
+#also, RADEX does not support uniform slab
 assert geometry != "uniform slab"
 
 
@@ -126,7 +120,7 @@ N_grid = np.logspace(13,18,grid_size)*constants.centi**-2
 Tkin_grid = np.logspace(np.log10(5),np.log10(500),grid_size)
 N_GRID,TKIN_GRID = np.meshgrid(N_grid,Tkin_grid,indexing='ij')
 
-nH2_grid = np.array((1e2,2e3,5e4,1e6))*constants.centi**-3
+nH2_grid = np.array((1e3,1e4,1e5))*constants.centi**-3
 residuals = {"Tex":[],"flux":[]}
 
 v = np.linspace(-width_v*3,width_v*3,30)
@@ -165,28 +159,34 @@ for i,nH2 in enumerate(nH2_grid):
                 flux[ID][i,j,k] = f
                 excitation_temp[ID][i,j,k] = Tex
 
-for quantity,values in zip(("flux","Tex"),(flux,excitation_temp)):
+nrows = nH2_grid.size
+ncols = 2
+fig,axes = plt.subplots(nrows=nrows,ncols=ncols,constrained_layout=True,
+                        figsize=(6,8))
+fig.suptitle(f"CO {plot_trans.name}")
+
+for column,(quantity,values) in enumerate(zip(("Tex","flux"),(excitation_temp,flux))):
     print(quantity)
-    fig,axes = plt.subplots(2,2,constrained_layout=True)
     for code,val in values.items():
         n_negative = (val<0).sum()
         print(f"{quantity} {code}: {n_negative}/{val.size} are negative")
-    residual = (values["pythonradex"]-values["radex"])/values["pythonradex"]
-    masked_residual = residual.copy()*100
-    assert -vmin == vmax
-    mask = np.abs(masked_residual) > vmax
-    masked_residual[mask] = np.nan
-    print(f"{quantity} with large difference between pythonradex and RADEX")
+        n_zero = (val==0).sum()
+        print(f"{quantity} {code}: {n_zero}/{val.size} are 0")
+    ratio = values["pythonradex"]/values["radex"]
+    #assert np.all(np.isfinite(ratio))
+    masked_ratio = ratio.copy()
[…]
