+++ joss_paper/code/benchmark_AndresMegias_original.py
@@ -1,0 +1,297 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+PythonRadex test
+
+Andrés
+"""
+
+import io
+import time
+import subprocess
+import numpy as np
+import pandas as pd
+import pythonradex as pradex
+from scipy import constants
+# import matplotlib.pyplot as plt
+
+radex_path = '/Users/andres/Applications/RADEX/bin/radex'
+
+def radex(molecule, freq_range, backgr_temp, kin_temp, molec_col_dens,
+          collider_num_dens, line_width):
+    """
+    Perform the RADEX calculation with the given parameters.
+
+    Parameters
+    ----------
+    molecule : str
+        RADEX name of the molecule.
+    freq_range : list (float)
+        Minimum and maximum frequencies to compute (GHz).
+    backgr_temp : str
+        Background temperature (K).
+    molec_kin_temp : str / array
+        Kinetic temperature (K).
+    molec_col_dens : str / array
+        Column density of the molecule (/cm2).
+    collider_num_dens : dict(float)
+        collider number densities (/cm3).
+    line_width : str
+        Width of the lines of the molecule (km/s).
+
+    Returns
+    -------
+    transitions_df : dataframe
+        List of calculated transitions by RADEX on-line
+    """
+    radex_name = species_abreviations[molecule]
+    def create_radex_input(molecule, min_freq, max_freq, kin_temp, backgr_temp,
+                           collider_num_dens, molec_col_dens, line_width):
+        """
+        Create the RADEX input file with the given parameters.
+        """
+        text = []
+        radex_name = species_abreviations[molecule]
+        if type(molec_col_dens) in (str, int, float, dict):
+            molec_col_dens = [molec_col_dens]
+        if type(collider_num_dens) in (str, int, float, dict):
+            collider_num_dens = [collider_num_dens]
+        if type(kin_temp) in (str, int, float, dict):
+            kin_temp = [kin_temp]
+        num_calcs = 0
+        for kin_temp_i in kin_temp:
+            for collider_num_dens_i in collider_num_dens:
+                for col_dens_i in molec_col_dens:
+                    num_collision_partners = len(collider_num_dens_i)
+                    text += [f'{radex_name}.dat']
+                    text += [f'{radex_name}.out']
+                    text += [f'{min_freq} {max_freq}']
+                    text += [f'{kin_temp_i}']
+                    text += [f'{num_collision_partners}']
+                    for (species,num_dens_i) in collider_num_dens_i.items():
+                        text += [f'{species}', f'{float(num_dens_i):e}']
+                    text += [f'{backgr_temp}']
+                    text += [f'{float(col_dens_i):e}']
+                    text += [f'{line_width}']
+                    text += ['1']
+                    num_calcs += 1
+        text[-1] = text[-1].replace('1', '0')
+        for i in range(len(text) - 1):
+            text[i] += '\n'
+        with open(f'{radex_name}.inp', 'w') as file:
+            file.writelines(text)
[…]

[…]
