### 2025-12-08 b60e15e
paper updates
