### 2025-12-08 61f3772
paper updates
