### 2025-12-08 eb8eab5
expose submodules
