+++ src/pythonradex/__init__.py
@@ -2,3 +2,5 @@
 """
 Python implementation of the RADEX radiative transfer code
 """
+
+from . import helpers, LAMDA_file, radiative_transfer
