+++ manual_tests_and_benchmarks/performance_tests/ng_acceleration_performance.py
@@ -38,19 +38,19 @@ timeit_number = 10
 for N in N_values:
     print(f'N={N/constants.centi**-2:.2g} cm-2')
     for ng in (True,False):
-        cloud = radiative_transfer.Cloud(
+        source = radiative_transfer.Source(
                     datafilepath=datafilepath,geometry=geometry,
                     line_profile_type=line_profile_type,width_v=width_v,
                     warn_negative_tau=False,use_Ng_acceleration=ng)
-        cloud.update_parameters(
+        source.update_parameters(
                      ext_background=ext_background,Tkin=Tkin,N=N,
                      collider_densities=collider_densities,T_dust=0,
                      tau_dust=0)
-        cloud.solve_radiative_transfer()
-        solve_time = timeit.timeit(cloud.solve_radiative_transfer,
+        source.solve_radiative_transfer()
+        solve_time = timeit.timeit(source.solve_radiative_transfer,
                                    number=timeit_number)
-        max_tau = np.max(cloud.tau_nu0_individual_transitions)
-        Tex_10 = cloud.Tex[0]
+        max_tau = np.max(source.tau_nu0_individual_transitions)
+        Tex_10 = source.Tex[0]
         print(f'time for ng {ng}, max tau {max_tau:.4g}, '
               +f'Tex={Tex_10:.4g}: {solve_time:.2g}')
     print('\n')
