+++ manual_tests_and_benchmarks/investigate_average_over_phi_nu/impact_of_averaging_for_dust.py
@@ -62,12 +62,12 @@ for test_case in test_cases:
         Tex = []
         taus = []
         for treat_overlap in (True,False):
-            cloud = radiative_transfer.Cloud(
+            source = radiative_transfer.Source(
                                   treat_line_overlap=treat_overlap,**cloud_kwargs)
-            cloud.update_parameters(**cloud_params)
-            cloud.solve_radiative_transfer()
-            Tex.append(cloud.Tex[trans_index])
-            taus.append(cloud.tau_nu0_individual_transitions[trans_index])
+            source.update_parameters(**cloud_params)
+            source.solve_radiative_transfer()
+            Tex.append(source.Tex[trans_index])
+            taus.append(source.tau_nu0_individual_transitions[trans_index])
         Tex_relative_diff = relative_diff(Tex)
         print(Tex_relative_diff)
         if Tex_relative_diff > 0.01:
