### 2025-12-11 3670beb
better line selection for spectrum
