+++ tests/test_flux.py
@@ -214,34 +214,140 @@ class TestInvalidFluxRequest():
                    fluxcalculator.fluxes_of_individual_transitions(solid_angle=1)
 
 
-def test_tau_nu_constructor():
+class TestVarious():
+
     width_v = 1*constants.kilo
     N = 1e12/constants.centi**2
     T = 344
-    geo_name = 'static sphere'
-    geo = radiative_transfer.Source.geometries[geo_name]
-    test_transitions = [0,3,10]
-    for lp in ('rectangular','Gaussian'):
-        mol = molecule.EmittingMolecule(datafilepath=CO_datafilepath,
-                                        line_profile_type=lp,width_v=width_v)
-        level_population = mol.LTE_level_pop(T)
-        tau_nu0 = mol.get_tau_nu0_lines(N=N,level_population=level_population)
-        fluxcalculator = flux.FluxCalculator(
-                                 emitting_molecule=mol,level_population=level_population,
-                                 geometry_name=geo_name,
-                                 compute_flux_nu=geo.compute_flux_nu,
-                                 tau_nu0_individual_transitions=tau_nu0,
-                                 tau_dust=zero,S_dust=zero,V_LVG_sphere=mol.width_v/2)
-        for t in test_transitions:
-            line = mol.rad_transitions[t]
-            width_nu = mol.width_v/constants.c * line.nu0
-            nu = np.linspace(line.nu0-2*width_nu,line.nu0+2*width_nu,200)
-            constructed_tau_nu = fluxcalculator.construct_tau_nu_individual_line(
-                                    line=line,nu=nu,tau_nu0=tau_nu0[t])
-            N1 = N*level_population[line.low.number]
-            N2 = N*level_population[line.up.number]
-            expected_tau_nu = line.tau_nu(N1=N1,N2=N2,nu=nu)
-            assert np.allclose(constructed_tau_nu,expected_tau_nu,atol=0,rtol=1e-3)
+
+    def generate_flux_calculator(self,geo_name,tau_dust,S_dust):
+        geo = radiative_transfer.Source.geometries[geo_name]
+        for lp in ('rectangular','Gaussian'):
+            mol = molecule.EmittingMolecule(datafilepath=CO_datafilepath,
+                                            line_profile_type=lp,width_v=self.width_v)
+            level_population = mol.LTE_level_pop(self.T)
+            tau_nu0 = mol.get_tau_nu0_lines(N=self.N,level_population=level_population)
+            fluxcalculator = flux.FluxCalculator(
+                                     emitting_molecule=mol,level_population=level_population,
+                                     geometry_name=geo_name,
+                                     compute_flux_nu=geo.compute_flux_nu,
+                                     tau_nu0_individual_transitions=tau_nu0,
+                                     tau_dust=tau_dust,S_dust=S_dust,V_LVG_sphere=mol.width_v/2)
+            yield {"mol":mol,"fluxcalculator":fluxcalculator,"tau_nu0":tau_nu0,
+                   "level_population":level_population}
+    
+    def test_tau_nu_constructor(self):
+        test_transitions = [0,3,10]
+        for f in self.generate_flux_calculator(geo_name="static sphere",tau_dust=zero,
+                                               S_dust=zero):
+            for t in test_transitions:
+                line = f["mol"].rad_transitions[t]
+                width_nu = f["mol"].width_v/constants.c * line.nu0
+                nu = np.linspace(line.nu0-2*width_nu,line.nu0+2*width_nu,200)
+                constructed_tau_nu = f["fluxcalculator"].construct_tau_nu_individual_line(
+                                        line=line,nu=nu,tau_nu0=f["tau_nu0"][t])
+                N1 = self.N*f["level_population"][line.low.number]
[…]
