+++ manual_tests_and_benchmarks/benchmarking_vs_RADEX_and_LIME/RADEX_vs_pythonradex_vs_LIME.py
@@ -198,8 +198,8 @@ for N_case,LTE_case in itertools.product(N_cases,coll_partner_density_cases):
     transition = source.emitting_molecule.rad_transitions[trans_number]
     up = transition.up
     low = transition.low
-    up_pop = level_pop[:,up.number]
-    low_pop = level_pop[:,low.number]
+    up_pop = level_pop[:,up.index]
+    low_pop = level_pop[:,low.index]
     Delta_E = transition.Delta_E
     lime_Tex = np.where(low_pop>0,-Delta_E/constants.k/np.log(low.g*up_pop/(up.g*low_pop)),
                         0)
@@ -212,7 +212,7 @@ for N_case,LTE_case in itertools.product(N_cases,coll_partner_density_cases):
         obs_flux_density[bb_key] = black_body_flux_density
         obs_flux[bb_key] = black_body_flux_density*width_nu
     elif N_case == 'thin' and LTE_case=='LTE':
-        up_level_pop = source.emitting_molecule.LTE_level_pop(Tkin)[up.number]
+        up_level_pop = source.emitting_molecule.LTE_level_pop(Tkin)[up.index]
         if general_geometry == 'slab':
             thin_LTE_flux = up_level_pop*N*slab_surface*transition.A21*Delta_E\
                              /(4*np.pi*slab_surface)*Omega
