+++ manual_tests_and_benchmarks/emerging_flux/check_sphere_flux.py
@@ -58,15 +58,15 @@ pythonradex_fluxes = {ID:np.empty(n_values.size) for ID in beta_funcs.keys()}
 beta_fluxes = {ID:np.empty(n_values.size) for ID in beta_funcs.keys()}
 
 for i,n in enumerate(n_values):
-    N1 = n*level_pop[trans.low.number]*2*r
-    N2 = n*level_pop[trans.up.number]*2*r
+    N1 = n*level_pop[trans.low.index]*2*r
+    N2 = n*level_pop[trans.up.index]*2*r
     tau_nu = atomic_transition.tau_nu(
                A21=trans.A21,phi_nu=phi_nu,
                g_low=trans.low.g,g_up=trans.up.g,N1=N1,N2=N2,nu=nu)
     source_func = helpers.B_nu(nu=nu,T=T)
     flux_kwargs = {'tau_nu':tau_nu,'source_function':source_func,'solid_angle':solid_angle}
     LVG_sphere_kwargs = {'nu':nu,'nu0':trans.nu0,'V':width_v/2}
-    flux_no_beta = volume*n*level_pop[trans.up.number]*trans.A21*trans.Delta_E\
+    flux_no_beta = volume*n*level_pop[trans.up.index]*trans.A21*trans.Delta_E\
                        /(4*np.pi*d**2) * phi_nu #W/m2/Hz
     if i == 0:
         assert np.max(tau_nu) < 0.01

[…]
