### 2025-12-11 46ec10e
improved line selection
