### 2025-12-12 058c1c7
working on brightness temperature
