+++ tests/test_radiative_transfer.py
@@ -5,7 +5,7 @@ Created on Wed Nov 15 10:30:44 2017
 @author: gianni
 """
 import os
-from pythonradex import radiative_transfer,helpers
+from pythonradex import radiative_transfer,helpers,escape_probability
 from scipy import constants
 import numpy as np
 import itertools
@@ -26,7 +26,7 @@ geometries = tuple(radiative_transfer.Source.geometries.keys())
 use_ng_options = (True,False)
 treat_overlap_options = (False,True)
 
-def get_general_test_cloud(specie,width_v):
+def get_general_test_source(specie,width_v):
     return radiative_transfer.Source(
                           datafilepath=datafilepath[specie],geometry='static sphere',
                           line_profile_type='Gaussian',width_v=width_v)
@@ -198,7 +198,7 @@ class TestUpdateParameters():
         for p_name in self.standard_params.keys():
             invalid_initial_params = self.standard_params.copy()
             invalid_initial_params[p_name] = None
-            source = get_general_test_cloud(specie='CO',width_v=1*constants.kilo)
+            source = get_general_test_source(specie='CO',width_v=1*constants.kilo)
             with pytest.raises(AssertionError):
                 source.update_parameters(**invalid_initial_params)
 
@@ -308,10 +308,10 @@ class TestUpdateParameters():
                                      original_params=current_params)
 
     def test_invalid_collider_densities(self):
-        test_cloud = get_general_test_cloud(specie='CO',width_v=1*constants.kilo)
+        test_source = get_general_test_source(specie='CO',width_v=1*constants.kilo)
         invalid_coll_densities = {'e':1/constants.centi**3} #CO has not electron data
         with pytest.raises(ValueError):
-            test_cloud.update_parameters(ext_background=cmb,
+            test_source.update_parameters(ext_background=cmb,
                                          N=1e12/constants.centi*2,Tkin=20,
                                          collider_densities=invalid_coll_densities)
 
@@ -411,7 +411,7 @@ def test_compute_residual():
 
 
 class TestModelGrid():
-    
+
     source = radiative_transfer.Source(
                           datafilepath=datafilepath['CO'],geometry='static sphere',
                           line_profile_type='rectangular',width_v=1.4*constants.kilo,
@@ -528,6 +528,9 @@ class TestModelGrid():
         for model in grid_with_failing_models:
             for rout in requested_output:
                 assert model[rout] is None
+
+    def test_brightness_temp(self):
+        raise NotImplementedError("need to implemente brightness temp support")
 
 
 def test_print_results():
@@ -667,4 +670,147 @@ def test_single_transition_molecule():
                 source.tau_nu(nu=nu)
                 source.spectrum(solid_angle=solid_angle,nu=nu)
                 source.print_results()
-    
+
+
+class TestBrightnessTemp():
+
+    N_thick = 1e20*constants.centi**-2
+    N_regular = 1e15*constants.centi**-2
+    N_thin = 1e12*constants.centi**-2
+    Tkin = 123
+    collider_densities = {"ortho-H2":1e5*constants.centi**-3,
+                          "para-H2":1e5*constants.centi**-3}
+    specie = "CO"
+    width_v = 1*constants.kilo
+    T_dust = 45
+    tau_dust_thick = 50
+    general_params = {"N":1e14*constants.centi**-2,"Tkin":101,
+                      "collider_densities":{"ortho-H2":1e5*constants.centi**-3},
+                      "ext_background":cmb,"T_dust":0,"tau_dust":0}
+    general_params_dust = {"N":1e14*constants.centi**-2,"Tkin":101,
+                           "collider_densities":{"ortho-H2":1e5*constants.centi**-3},
+                           "ext_background":cmb,
+                           "T_dust":lambda nu: np.ones_like(nu)*78,
[…]
