+++ tests/test_flux.py
@@ -366,6 +366,114 @@ class TestVarious():
                 assert single_spec[0] == multi_spec[1]
 
 
+class TestIntensityNu0NoOverlap():
+
+    Tkin = 100
+
+    def generate_flux_calculator(self,geometry_name,line_profile_type,
+                                 S_dust,tau_dust,width_v,datafilepath):
+        mol = molecule.EmittingMolecule(
+                datafilepath=datafilepath,line_profile_type=line_profile_type,
+                width_v=width_v)
+        level_population = mol.LTE_level_pop(T=self.Tkin)
+        geometry = radiative_transfer.Source.geometries[geometry_name]
+        tau_nu0_individual_transitions = np.arange(mol.n_rad_transitions,dtype=float)+0.1
+        tau_nu0_individual_transitions = tau_nu0_individual_transitions[::-1]
+        tau_nu0_individual_transitions /= np.max(tau_nu0_individual_transitions)
+        fluxcalculator = flux.FluxCalculator(
+                           emitting_molecule=mol,level_population=level_population,
+                           geometry_name=geometry_name,V_LVG_sphere=width_v/2,
+                           compute_flux_nu=geometry.compute_flux_nu,
+                           tau_nu0_individual_transitions=tau_nu0_individual_transitions,
+                           tau_dust=tau_dust,S_dust=S_dust)
+        return fluxcalculator
+
+    def test_intensity_nu0(self):
+        for geo,line_profile_type in\
+                itertools.product(radiative_transfer.Source.geometries.keys(),
+                                  radiative_transfer.Source.line_profiles.keys()):
+            if "LVG" in geo and line_profile_type == "Gaussian":
+                continue
+            kwargs = {"no dust":{"geometry_name":geo,"line_profile_type":line_profile_type,
+                                 "S_dust":zero,"tau_dust":zero,"width_v":1*constants.kilo,
+                                 "datafilepath":CO_datafilepath}}
+            if "LVG" not in geo:
+                kwargs["dust"] = kwargs["no dust"].copy()
+                kwargs["dust"]["S_dust"] = lambda nu: helpers.B_nu(nu=nu,T=200)
+                kwargs["dust"]["tau_dust"] = lambda nu: np.ones_like(nu)*0.5
+            for kw in kwargs.values():
+                f = self.generate_flux_calculator(**kw)
+                transitions = np.arange(f.emitting_molecule.n_rad_transitions)
+                Omega = 0.235
+                intensity_nu0 = f.intensity_nu0_no_overlap(transitions=transitions)
+                nu = f.emitting_molecule.nu0
+                f.set_nu(nu=nu)
+                expected_from_spec = f.spectrum(solid_angle=Omega)/Omega
+                assert np.allclose(intensity_nu0,expected_from_spec,atol=0,rtol=1e-3)
+                #now explicit calculation:
+                tau_dust = kw["tau_dust"](nu)
+                S_dust = kw["S_dust"](nu)
+                tau_tot = f.tau_nu0_individual_transitions+tau_dust
+                S_lines = helpers.B_nu(nu=nu,T=f.Tex)
+                source_function = np.where(tau_tot==0,0,
+                                           ( f.tau_nu0_individual_transitions*S_lines+tau_dust*S_dust)/tau_tot)
+                flux_kwargs = {"tau_nu":tau_tot,"source_function":source_function,
+                               "solid_angle":Omega}
+                if "LVG" in geo:
+                    explicit = escape_probability.compute_flux_nu0_lvg_sphere(**flux_kwargs)/Omega
+                else:
+                    explicit = f.compute_flux_nu(**flux_kwargs)/Omega
[…]
