### 2026-01-07 b8035c8
intensity updates
