+++ src/pythonradex/helpers.py
@@ -70,9 +70,29 @@ def Delta_nu(Delta_v,nu0):
 
 @nb.jit(nopython=True,cache=True)
 def RJ_brightness_temperature(specific_intensity,nu):
+    r'''calculates the Rayleigh-Jeans brightness temperature
+    
+    Args:
+        specific_intensity (:obj:`float` or numpy.ndarray): the specific intensity in
+        [W/m\ :sup:`2`/Hz/sr]
+        nu (:obj:`float` or numpy.ndarray): the frequency in [Hz]
+
+    Returns:
+        float or numpy.ndarray: the Rayleigh-Jeans brightness temperature in [K]
+        '''
     return specific_intensity*constants.c**2/(2*nu**2*constants.k)
 
 @nb.jit(nopython=True,cache=True)
 def Planck_brightness_temperature(specific_intensity,nu):
+    r'''calculates the Planck brightness temperature
+    
+    Args:
+        specific_intensity (:obj:`float` or numpy.ndarray): the specific intensity in
+        [W/m\ :sup:`2`/Hz/sr]
+        nu (:obj:`float` or numpy.ndarray): the frequency in [Hz]
+
+    Returns:
+        float or numpy.ndarray: the Planck brightness temperature in [K]
+        '''
     log_term = np.log(2*constants.h*nu**3/(constants.c**2*specific_intensity)+1)
     return constants.h*nu/constants.k*log_term**-1
