+++ manual_tests_and_benchmarks/performance_tests/pythonradex_vs_radex/radex_vs_pythonradex_performance.py
@@ -23,7 +23,10 @@ import mini_radex_wrapper
 
 
 ext_background = helpers.generate_CMB_background(z=0)
-n_elements = [5,7,10,15,20] #for Tkin, collider and N
+
+#n_elements = [5,7,10,15,20] #for Tkin, collider and N
+n_elements = [5,7,10]
+
 #most optimistic case is if only N is varied because then rate equations don't
 #need to be re-calculated every time
 #however, seems like it doesn't really change anything...
@@ -34,10 +37,10 @@ vary_only_N = False
 geometry = 'static sphere'
 # geometry = 'LVG slab'
 
-data_filename = 'co.dat'
-colliders = ['para-H2','ortho-H2']
-log_N_limits = 13+4,18+4
-Tmin,Tmax = 20,250
+# data_filename = 'co.dat'
+# colliders = ['para-H2','ortho-H2']
+# log_N_limits = 13+4,18+4
+# Tmin,Tmax = 20,250
 #narrow range where RADEX does not throw warnings:
 # log_N_limits = 16+4,16.3+4
 # Tmin,Tmax = 50,51
@@ -52,19 +55,21 @@ Tmin,Tmax = 20,250
 # log_N_limits = 10+4,12+4
 # Tmin,Tmax = 60,250
 
-# data_filename = 'c.dat'
-# colliders = ['para-H2','ortho-H2']
-# log_N_limits = 12+4,18+4
-# Tmin,Tmax = 60,250
+data_filename = 'c.dat'
+colliders = ['para-H2','ortho-H2']
+log_N_limits = 12+4,18+4
+Tmin,Tmax = 60,250
 
 
-# #ATTENTION: if no H2 is given, RADEX just puts 1e5 cm-3 by default! WTF!
+#ATTENTION: if no H2 is given, RADEX just puts 1e5 cm-3 by default! WTF!
 #RADEX does other strange things with the H2 density:
 #see line 112 in io.f, line 168 in io.f, and line 225 in readdata.f
 for collider in colliders:
     assert 'H2' in collider
 
-line_profile_type = 'rectangular' #actually, RADEX assumes rectangular, but than converts it Gaussian for the line flux
+#actually, RADEX assumes rectangular, but than converts it Gaussian for optical
+#depth and the line flux
+line_profile_type = 'rectangular' 
 width_v = 1*constants.kilo
 use_Ng_acceleration = True
 treat_line_overlap = False
@@ -120,7 +125,8 @@ for i,n in enumerate(n_elements):
                                 collider_densities=collider_densities,N=N,T_dust=0,
                                 tau_dust=0)
         source.solve_radiative_transfer()
-        source.fluxes_of_individual_transitions(solid_angle=1,transitions=None)
+        source.frequency_integrated_emission_of_individual_transitions(
+                     output_type="flux",solid_angle=1,transitions=None)
     end = time.time()
     pythonradex_times[i] = end-start
 
@@ -129,16 +135,20 @@ for i,n in enumerate(n_elements):
         remove_pythonradex_cache()
     start = time.time()
     source = radiative_transfer.Source(**cloud_kwargs)
-    requested_output=['level_pop','Tex','tau_nu0_individual_transitions',
-                      'fluxes_of_individual_transitions']
     collider_densities_values={collider:coll_density_values for collider in
                                colliders}
-    grid = source.model_grid(ext_backgrounds={'extbg':ext_background},N_values=N_values,
-                            Tkin_values=Tkin_values,
-                            collider_densities_values=collider_densities_values,
-                            requested_output=requested_output,solid_angle=1)
-    for model in grid:
-        pass
+    iterator = source.efficient_parameter_iterator(
+                 ext_backgrounds={'extbg':ext_background},
[…]
