### 2026-01-14 aeac74e
notebook updates; removed grid iterator
