+++ tests/test_radiative_transfer.py
@@ -419,11 +419,6 @@ def test_intensity_transformation():
         radiative_transfer.Source.transform_specific_intensity(
                    specific_intensity=specific_intensity_Planck,
                    nu=nu,solid_angle=None,output_type="flux density")
-    for output_type in ("Rayleigh-Jeans","Planck"):
-        with pytest.warns(UserWarning):
-            radiative_transfer.Source.transform_specific_intensity(
-                       specific_intensity=specific_intensity_Planck,nu=nu,
-                       solid_angle=1,output_type=output_type)
     with pytest.raises(ValueError):
         radiative_transfer.Source.transform_specific_intensity(
                    specific_intensity=specific_intensity_Planck,nu=nu,
@@ -436,6 +431,26 @@ def test_intensity_transformation():
                output_type="Planck")
     for brightness_T in (T_RJ,T_Planck):
         assert np.allclose(brightness_T,T,atol=0,rtol=1e-8)
+
+def test_solid_angle_warnings():
+    source = get_general_test_source(specie="CO", width_v=1.3*constants.kilo)
+    source.update_parameters(N=1e15*constants.centi**-2,Tkin=123,
+                             collider_densities={"ortho-H2":1e6*constants.centi**-3},
+                             ext_background=cmb,T_dust=0,tau_dust=0)
+    source.solve_radiative_transfer()
+    with pytest.warns(UserWarning):
+        source.frequency_integrated_emission(
+                        output_type="intensity",transitions=None,solid_angle=1.23)
+    for output_type in ("specific intensity","Rayleigh-Jeans","Planck"):
+        with pytest.warns(UserWarning):
+            trans = source.emitting_molecule.rad_transitions[1]
+            width_v = trans.line_profile.width_v
+            v = np.linspace(-2*width_v,2*width_v,20)
+            nu = trans.nu0*(1-v/constants.c)
+            source.spectrum(nu=nu, output_type=output_type,solid_angle=0.23)
+        with pytest.warns(UserWarning):
+            source.emission_at_line_center(output_type=output_type,transitions=None,
+                                           solid_angle=0.45)
 
 def CO_HCl_model_iterator():
     use_Ng_acceleration = True
@@ -470,12 +485,12 @@ spec_output_types = ("specific intensity","flux density","Rayleigh-Jeans","Planc
 @pytest.mark.filterwarnings("ignore:LVG sphere geometry")
 def test_emission_at_line_center():
     for source in CO_HCl_model_iterator():
-        for transitions in (None,[0,2],[1,]):
+        for transitions in (None,[0,2],1):
             if transitions is None:
                 nu0s = source.emitting_molecule.nu0
             else:
                 nu0s =  np.array([source.emitting_molecule.rad_transitions[i].nu0
-                                  for i in transitions])
+                                  for i in np.atleast_1d(transitions)])
             for output_type in spec_output_types:
                 solid_angle = 0.23 if output_type=="flux density" else None
                 emission = source.emission_at_line_center(
@@ -485,6 +500,59 @@ def test_emission_at_line_center():
                                       nu=nu0s,output_type=output_type,
                                       solid_angle=solid_angle)
                 assert np.allclose(emission,expected_emission,atol=0,rtol=1e-3)
+
+def test_transition_index_transformation():
+    source = get_general_test_source(specie="CO", width_v=1.23*constants.kilo)
+    None_indices = source.transform_transition_indices(indices=None)
+    assert np.all(None_indices==np.arange(source.emitting_molecule.n_rad_transitions))
+    indices_0d = source.transform_transition_indices(indices=2)
+    assert indices_0d.ndim == 1
+    assert len(indices_0d) == 1
+    assert indices_0d[0] == 2
[…]
