+++ manual_tests_and_benchmarks/performance_tests/pythonradex_vs_radex/performance_comparison.py
@@ -97,7 +97,7 @@ for grid_name,grid in grids.items():
                                         collider_densities=collider_densities,N=N,T_dust=0,
                                         tau_dust=0)
                 source.solve_radiative_transfer()
-                source.frequency_integrated_emission_of_individual_transitions(
+                source.frequency_integrated_emission(
                            output_type="flux",solid_angle=1,transitions=None)
             end = time.time()
             pythonradex_time = end-start
