+++ joss_paper/code/molpop-cep-master/DataBase/lamda.f90
@@ -1,399 +1,0 @@
-
-Module File_Units
-   character(80)    :: Lname, inp_file, out_file
-   integer          :: inunit = 10, outunit = 11, lunit = 12
-End Module File_Units
-
-Module Mol_Info
-   character(40)    :: Mol_name
-   integer          :: Nlev
-   real             :: weight
-End Module Mol_Info
-
-!***********************************************************************
-
-Program Leiden
-
-! Convert atomic data files from the Leiden database to MOLPOP input format
-! Files to be converted can be listed in a file (one per line) and the program invoked with 
-!   the (arbitrary) filename as an argument
-! If the program is invoked without an argument, it prompts for names of files to convert
-! The Leiden file names can be specified with or without their .dat extension
-! A Leiden file 'fname.dat' will produce 'fname_lamda.molecule' with the energy levels 
-!  and A-coefficients in MOLPOP format, and a set of 'Coll/fname_collider_lamda.kij' 
-!  for the collision rate coefficients of each collider in the Leiden lamda file 
-
-  use File_Units
-  Implicit None
-  integer          :: Open_Status
-  character(80)    :: File_List
-
-  print "(/,'  ****  Leiden-to-MOLPOP Conversion ****',/)"
-
-! Now work:
-  call GETARG(1, File_List)
-  if (trim(File_List).eq."") then ! use keyboard entry
-     call Keyboard_entry
-  else                            ! work on list from the file whose name was entered as argument
-     call File_list_entry(trim(File_List))
-  end if   
-
-  print*, 'Done!'
-  stop
-End Program Leiden
-
-!***********************************************************************
-
-Subroutine Keyboard_entry
-
-! Keyboard entry of Leiden input filename , with or without the .dat extension
-
-  IMPLICIT None
-  character(80) :: str
-
-  do  
-    WRITE(*,'("Enter a Leiden database file name (q to quit): ")', ADVANCE='NO')
-    !WRITE(*,'(": ")', ADVANCE='NO')   
-    read *, str
-    if (str == 'q') return ! done; otherwise
-    call Process(str)      ! work on the file entered and get another one
-  end do
-  
-End Subroutine Keyboard_entry
-
-!***********************************************************************
-
-Subroutine File_list_entry(fname)
-
-! Entry of Leiden input filename, with or without the .dat extension
-! from a list of files in 'fname' 
-
-   use File_Units
-   IMPLICIT None
-   character(*), intent(in) :: fname
-   character(80) :: datfile
-   integer       :: Open_Status, line_status
-
-   open(lunit, file=fname, status='Old', action='Read', IOSTAT = Open_Status)
-   if (Open_Status > 0) then
-       print *, '***** Error; file ',fname,' did not open properly *****' 
-       return
-   end if
-   line_status = 0
-   do
-      read(lunit,'(a)',IOSTAT = line_status) datfile
-      if (line_status /= 0) return          ! done, end of the file; otherwise,
-      call Process(trim(datfile))           ! work on the next entry in the file list
-   end do
-   
-End Subroutine File_list_entry
-
-!***********************************************************************
-
-Subroutine Process(str)
-! This is the centarl control for processing the 
-! Leiden database file 'Lname.dat' into MOLPOP data files 
-
-   use File_Units
-   use Mol_Info
-   IMPLICIT None
-   integer          :: ind, Open_Status
-   character(*), intent(in) :: str
-    
-    ind = index(str,'.dat')
-    if (ind == 0) then ! the file name was specified without the .dat extension
-        Lname    = str
-        inp_file = trim(Lname)//'.dat'
[…]
