+++ joss_paper/code/molpop-cep-master/Source/maths_molpop.f90
@@ -1,1905 +1,0 @@
-! Contains all the mathematical routines and other routines that cope
-! with characters
-module maths_molpop
-use global_molpop
-implicit none
-contains
-
-
-      SUBROUTINE Pass_Header(iunit)
-!     Get past Header lines of opened file on unit # iunit
-!     The end of headers is marked by a line that contains just ">"
-      integer iunit
-      character*128 header
-      header = "XXX"
-      do while(.not. (header(1:1).eq.">"))
-         read(iunit, '(a)') header
-      end do
-      return
-      end subroutine Pass_Header
-
-
-      function error_message(opt,str)
-      character(len=128) :: opt, str
-      integer :: iunit
-      logical :: error_message
-
-      do iunit = 6,16,10
-            WRITE (IUNIT,"(/,'*** Terminated because of INPUT error ***',/, &
-            '    Wrong option for ',a,':  ',a)") trim(adjustl(opt)), trim(adjustl(str))
-      end do
-      error_message = .true.
-      return
-      end function error_message
-
-
-      SUBROUTINE OPTDEP(X)
-!     CALCULATE THE OPTICAL DEPTHS AND ESCAPE PROBABILITIES
-      use global_molpop
-      INTEGER I,J
-      DOUBLE PRECISION X(N), TAUd
-
-      IF (R.EQ.0.) THEN
-         DO I = 2, N
-           DO J = 1, I - 1
-              TAU(I,J) = 0.
-              ESC(I,J) = 1.
-           END DO
-         END DO
-         RETURN
-      END IF
-!
-      DO I = 2, N
-        DO J = 1, I - 1
-          IF (TAUX(I,J).NE.0.) THEN
-            TAU(I,J) = R*TAUX(I,J)*(X(J)*GAP(I,J)-X(I))
-            if (dustAbsorption) then
-               TAUd     = R*qdust(I,J)
-               TAU(I,J) = TAU(I,J) + TAUd
-               Xd(I,J)  = TAUd/TAU(I,J)
-            end if
-            IF (KBETA.EQ.2) THEN
-               CALL BETA_SLAB(TAU(I,J),ESC(I,J),DBDTAU(I,J))
-            ELSEIF (KBETA.EQ.1) THEN
-               CALL BETA_SPHERE(TAU(I,J),ESC(I,J),DBDTAU(I,J))
-            ELSEIF (KBETA.EQ.0) THEN
-               CALL BETA_LVG(TAU(I,J),ESC(I,J),DBDTAU(I,J))
-            ELSEIF (KBETA.EQ.-1) THEN
-               CALL BETA_LVG(3.0*TAU(I,J),ESC(I,J),DBDTAU(I,J))
-               DBDTAU(I,J) = 3.0 * DBDTAU(I,J)
-            END IF
-          END IF
-        END DO
-      END DO
-
-      RETURN
-      END SUBROUTINE OPTDEP
-
-      SUBROUTINE BETA_SLAB(TAUIN,BETA,DBDX)
-!     SLAB ESCAPE ESCAPE PROBABILITY AND ITS DERIVATIVE
-!     WITH RESPECT TO TAU
-!     BASED ON KROLIK & MCKEE 1978, ApJS 37, 459; eq 11
-!     small- and large-tau expressions are joined at tau = 3.41 instead of 5
[…]
