### 2026-01-15 0a9f8ed
added discussion about different H2 behaviour of RADEX and pythonradex to the docs
