### 2026-01-23 78f28b9
pythonradex vs RADEX paper figure
