+++ joss_paper/code/compare_pythonradex_RADEX.py
@@ -15,21 +15,21 @@ import os
 import matplotlib.pyplot as plt
 import numpy as np
 
-save_figure = False
+save_figure = True
 
 data_filename = "co.dat"
-#can only use LVG slab and uniform sphere
-geometry = "LVG slab"
-geometry_radex = "LVG slab"
-# geometry = "static sphere"
-# geometry_radex = "static sphere"
+#can only use static sphere
+#(for LVG sphere, RADEX uses different escape prob.; RADEX does not support
+#static slab; LVG slab needs rectangular profile)
+geometry = "static sphere"
+geometry_radex = "static sphere"
 
 #for grid comparison:
 plot_trans_index = 1
-grid_size = 25
-#grid_size = 5
+#grid_size = 25
+grid_size = 10
 
-vmin,vmax = 0.95,1.05
+vmin,vmax = 0.998,1.002
 cmap = plt.get_cmap("PuOr").copy()
 cmap.set_bad("red")
 
@@ -39,37 +39,41 @@ cmap.set_bad("red")
 #for consistency with RADEX:
 T_background = 2.73
 ext_background = helpers.generate_CMB_background()
-line_profile_type = "rectangular"
+#RADEX computes using rectangular, but uses a correction factor to convert
+#to Gaussian for flux and optical depth
+line_profile_type = "Gaussian"
 
-solid_angle = 1
 width_v = 1*constants.kilo
 datafilepath = os.path.join("../../tests/LAMDA_files",data_filename)
 
 #don't use LVG sphere because RADEX uses a different formula for the escape
 #probability
 assert geometry != "LVG sphere" and geometry_radex != "LVG sphere"
-#also, RADEX does not support uniform slab
-assert geometry != "uniform slab"
+#also, RADEX does not support static slab
+assert geometry != "static slab"
+#LVG slab needs rectangular profile
+assert geometry != "LVG slab"
 
 
 source = radiative_transfer.Source(datafilepath=datafilepath,geometry=geometry,
                                    line_profile_type=line_profile_type,
                                    width_v=width_v)
+#set ext background, dust:
+source.update_parameters(ext_background=ext_background,N=1e15/constants.centi**2,
+                         Tkin=94,collider_densities={"ortho-H2":1},
+                         T_dust=0,tau_dust=0)
 def compute_pythonradex_model(N,Tkin,coll_partner_densities):
-    source.update_parameters(ext_background=ext_background,N=N,Tkin=Tkin,
-                             collider_densities=coll_partner_densities,
-                             T_dust=0,tau_dust=0)
+    source.update_parameters(N=N,Tkin=Tkin,
+                             collider_densities=coll_partner_densities)
     source.solve_radiative_transfer()
-    pop_up = [source.level_pop[t.up.index] for t in
-              source.emitting_molecule.rad_transitions]
-    pop_low = [source.level_pop[t.low.index] for t in
-               source.emitting_molecule.rad_transitions]
-    return source.Tex,pop_up,pop_low
+    return source.Tex,source.upper_level_population,source.lower_level_population,\
+            source.tau_nu0_individual_transitions
 
 def compute_RADEX_model(N,Tkin,coll_partner_densities,transitions):
     Tex = []
     pop_up = []
     pop_low = []
+    tau = []
     epsilon = 1*constants.giga
     for trans in transitions:
[…]
