### 2026-01-23 22c2f70
paper updates
