### 2026-01-23 #9233 @gica3618
Dear all, I updated Fig. 1 of the paper as follows:

- Now comparing optical depth instead of flux in the second row, thank you @andresmegias for the suggestion. For flux or brightness temperature, RADEX performs a background subtraction, so some additional conversions are necessary to compare the RADEX output to pythonradex. For optical depth, that is not necessary, so it is easier to compare. Indeed, excitation temperature and optical depth are the most fundamental outputs of the radiative transfer calculation, so I think it makes sense to benchmark with those two quantities.
- Previously I ran pythonradex with a rectangular line profile for the comparison, but since RADEX applies a conversion factor to convert the rectangular optical depth profile to Gaussian, it is actually more appropriate to run pythonradex with a Gaussian line profile. With this change, the agreement between the two codes is now very good, mostly the difference is much smaller than 1%.
- (minor) instead of showing column density vs kinetic temperature for fixed H2 density, I now show H2 density vs column density for fixed kinetic temperature.

Please have a look at the updated figure.

Next, I will work on the other points raised by @andresmegias. I am not sure about the timeline, but please allow another two weeks or so.
