### 2026-01-23 #9233 @gica3618
Addressing the comments made by @andresmegias above:

> I'm also curious about this correction factor that Radex uses for computing the flux

> I asked about this in my last comment but I've already understood it. This is formally correct if the line is optically thin, right?

I think for optical depth it is always correct, but for flux only if the line is optically thin.

> Indeed, my main comment about PythonRadex is related to the last point: the different formulae used for computing the intensity / Rayleigh-Jeans brightness temperature for spheric geometries with respect to Radex. I'm not an expert on radiative transfer, but I'm trying to understand the formalism to assess if this different formula and the subsequent correction of the 1.5 factor (for optically thin lines) is indeed correct. On the one hand, I think the explanations given in the paper and especially in the documentation make sense, but on the other hand, I'm surprised that no one noticed this before in all these almost 20 years (especially Floris van der Tak, the author of Radex), also given that the static sphere is the default geometry in Radex. Have you discussed this with Floris himself? I think it would be valuable to know his opinion.

Let me make some more comments on this point:

- I believe that if we are talking about total flux (W/m2) emitted by the sphere, then pythonradex gives the correct answer. This can indeed be demonstrated by considering the optically thin limit where all photons escape, as described in the paper. On the other hand, note that brightness temperature (or specific intensity) depends on the position on the sphere. I mean, if you look towards the center of the sphere, it will be brighter than the sphere edges (because when looking through the center, the column density is higher), unless the sphere is completely optically thick. The brightness temperature computed by RADEX corresponds to the brightness temperature one would observe towards the center of the sphere if the observations resolved the sphere. On the other hand, pythonradex computes a sort of mean brightness temperature, in the sense that if you multiply that brightness temperature by the solid angle of the sphere, you get the correct total flux. So the brightness temperature computed by pythonradex corresponds to the brightness temperature one would measure from unresolved observations. And again, whether resolved or unresolved, pythonradex gives you the correct total flux.
- I emailed Floris in 2019 about this. He replied that he was not sure about it, and asked whether I could make a comparison between the two codes. So I emailed a figure showing the flux of RADEX and pythonradex as function of optical depth. I also sent him a pdf of the original derivation in the Osterbrock book. But I never got any answer anymore to this second email.
- Personally, when I think about all the other simplifications these codes make (assuming a uniform sphere, which is never true in reality; making the escape probability approximation; etc), the factor 1.5 is probably not that dramatic, although it's nice to have it corrected.

> Regarding the formal explanations of the use of one formula or another, I think it would be necessary to have this well explained either in the PythonRadex paper or in the documentation. Since there's a size limit for the paper, in the documentation it's perfectly fine. I've seen that @gica3618 has indeed added in the Cloud geometry section the mathematical derivation of the formulae for the escape probability and flux of the spherical geometries (static and LVG). This is greatly appreciated, since the original cited references were books that are not freely available on the internet.

I believe this is addressed by the updates I did to the documentation a while ago, but let me know if not.

> I still don't quite get what is the physical radius of the sphere (just in the static sphere, for example), because when running Radex (or PythonRadex) we don't input any radius or size for the sphere. Is it computed automatically so that the observed intensity / R-J brightness temperature corresponds to the input molecular column density?

That's a good question I never thought about. Indeed, the size of the sphere does not enter the radiative transfer calculation. And there is also no radius calculated automatically. The point is, the escape probability only depends on optical depth, not on the size of the sphere. What happens internally is that from the optical depth, the escape probability is computed, which then gives a sort of mean radiation field in the sphere. This field is used to solve for the level populations, which gives an updated optical depth. The size of the sphere never enters the calculation.

> This is very minor, but it would be appreciated if, for the escape probability of the static slab, you put the final formula resolving the integral (it's in the Radex paper, I think); and, similarly, when introducing the formulae for the escape probabilities of the LVG sphere and LVG slab, you could cite the exact equation number for each of the cited references. (Ideally it would be great if the derivation of those formulae were also present in the documentation, just for the sake of completeness; but for the review I think it's not necessary.)

For the static slab, I don't think the integral has an analytical solution. And it's not in the RADEX paper, since RADEX does not support the static slab geometry. For LVG sphere and LVG slab, I added the exact equation number or section in the cited paper.

> Lastly, I would suggest reviewing the documentation (especially in Differences between pythonradex and Radex and Cloud geometries) for making a clearer distinction between spectral flux density (Fν) and just flux (F, i.e., spectral flux density integrated over the line profile), since in some cases both terms are used kind of interchangeably and it might be a bit confusing for the reader. Similarly and for the sake of completeness, you could remind to the reader somewhere the relation between the intensity (Iν) and the spectral flux density: dFν = Iν·dΩ (Ω: solid angle).

This is a good point. I now made an effort to clarify the terminology both in the documentation, the code and the example notebooks. I settled on the following terminology:

- W/m2/Hz/sr: specific intensity
- W/m2/sr: intensity
- W/m2/Hz: flux density
- W/m2: flux

Finally, I put the updated documentation at a different url: https://pythonradex.readthedocs.io/en/joss_paper/. Please visit that link if you want to check the updates. This is to avoid confusion for users that are using the last release of the code, which is not aligned anymore with the updated documentation. Once I address the remaining comments and the paper is accepted, I will make a new release and the updated documentation will become official.
