### 2026-01-23 #9233 @gica3618
Just mentioning another update in the documentation, in the Section "Differences between pythonradex and RADEX", I added a section describing that RADEX adds a default H2 density whenever the LAMDA file contains H2 as a collider and the user does not specify a H2 density. Besides being a somewhat surprising behaviour, this makes it effectively impossible to run a RADEX model with the H2 density equal to 0. If explicitly set to 0, RADEX will not accept it (but it will accept a value of 1e-3 cm-3, which is very small).
