+++ joss_paper/code/compare_pythonradex_RADEX.py
@@ -7,10 +7,11 @@ Created on Fri Nov 14 13:51:56 2025
 """
 
 import sys
-sys.path.append('../../manual_tests_and_benchmarks/RADEX_wrapper')
+
+sys.path.append("../../manual_tests_and_benchmarks/RADEX_wrapper")
 import radex_wrapper
 from scipy import constants
-from pythonradex import radiative_transfer,helpers
+from pythonradex import radiative_transfer, helpers
 import os
 import matplotlib.pyplot as plt
 import numpy as np
@@ -18,97 +19,138 @@ import numpy as np
 save_figure = True
 
 data_filename = "co.dat"
-#can only use static sphere
-#(for LVG sphere, RADEX uses different escape prob.; RADEX does not support
-#static slab; LVG slab needs rectangular profile)
+# can only use static sphere
+# (for LVG sphere, RADEX uses different escape prob.; RADEX does not support
+# static slab; LVG slab needs rectangular profile)
 geometry = "static sphere"
 geometry_radex = "static sphere"
 
-#for grid comparison:
+# for grid comparison:
 plot_trans_index = 1
 grid_size = 10
 
-vmin,vmax = 0.998,1.002
+vmin, vmax = 0.998, 1.002
 cmap = plt.get_cmap("PuOr").copy()
 cmap.set_bad("red")
 
 
 ##############################################################
 
-#for consistency with RADEX:
+# for consistency with RADEX:
 T_background = 2.73
 ext_background = helpers.generate_CMB_background()
-#RADEX computes using rectangular, but uses a correction factor to convert
-#to Gaussian for flux and optical depth
+# RADEX computes using rectangular, but uses a correction factor to convert
+# to Gaussian for flux and optical depth
 line_profile_type = "Gaussian"
 
-width_v = 1*constants.kilo
-datafilepath = os.path.join("../../tests/LAMDA_files",data_filename)
+width_v = 1 * constants.kilo
+datafilepath = os.path.join("../../tests/LAMDA_files", data_filename)
 
-#don't use LVG sphere because RADEX uses a different formula for the escape
-#probability
+# don't use LVG sphere because RADEX uses a different formula for the escape
+# probability
 assert geometry != "LVG sphere" and geometry_radex != "LVG sphere"
-#also, RADEX does not support static slab
+# also, RADEX does not support static slab
 assert geometry != "static slab"
-#LVG slab needs rectangular profile
+# LVG slab needs rectangular profile
 assert geometry != "LVG slab"
 
 
-source = radiative_transfer.Source(datafilepath=datafilepath,geometry=geometry,
-                                   line_profile_type=line_profile_type,
-                                   width_v=width_v)
-#set ext background, dust:
-source.update_parameters(ext_background=ext_background,N=1e15/constants.centi**2,
-                         Tkin=94,collider_densities={"ortho-H2":1},
-                         T_dust=0,tau_dust=0)
-def compute_pythonradex_model(N,Tkin,coll_partner_densities):
-    source.update_parameters(N=N,Tkin=Tkin,
-                             collider_densities=coll_partner_densities)
-    source.solve_radiative_transfer()
-    return source.Tex,source.upper_level_population,source.lower_level_population,\
-            source.tau_nu0_individual_transitions
+source = radiative_transfer.Source(
+    datafilepath=datafilepath,
+    geometry=geometry,
+    line_profile_type=line_profile_type,
[…]
