+++ src/pythonradex/atomic_transition.py
@@ -11,84 +11,89 @@ from scipy import constants
 import numba as nb
 
 
-@nb.jit(nopython=True,cache=True) 
-def tau_nu(A21,phi_nu,g_low,g_up,N1,N2,nu):
-    return constants.c**2/(8*np.pi*nu**2)*A21*phi_nu*(g_up/g_low*N1-N2)
-
-@nb.jit(nopython=True,cache=True,error_model='numpy')
-def Tex(Delta_E,g_low,g_up,x1,x2):
-    return np.where((x1==0) & (x2==0),0,
-                    -Delta_E/(constants.k*np.log(g_low*x2/(g_up*x1))))
-
-@nb.jit(nopython=True,cache=True)
-def compute_K12(K21,g_up,g_low,Delta_E,Tkin):
-    #see RADEX manual
-    return (g_up/g_low*K21*np.exp(-Delta_E/(constants.k*Tkin)))
+@nb.jit(nopython=True, cache=True)
+def tau_nu(A21, phi_nu, g_low, g_up, N1, N2, nu):
+    return (
+        constants.c**2 / (8 * np.pi * nu**2) * A21 * phi_nu * (g_up / g_low * N1 - N2)
+    )
 
 
-class LineProfile():
-    
-    '''Abstract class representing a general, normalised line profile.'''
+@nb.jit(nopython=True, cache=True, error_model="numpy")
+def Tex(Delta_E, g_low, g_up, x1, x2):
+    return np.where(
+        (x1 == 0) & (x2 == 0),
+        0,
+        -Delta_E / (constants.k * np.log(g_low * x2 / (g_up * x1))),
+    )
+
+
+@nb.jit(nopython=True, cache=True)
+def compute_K12(K21, g_up, g_low, Delta_E, Tkin):
+    # see RADEX manual
+    return g_up / g_low * K21 * np.exp(-Delta_E / (constants.k * Tkin))
+
+
+class LineProfile:
+    """Abstract class representing a general, normalised line profile."""
 
     n_nu_elements_for_coarse_array = 10
 
-    def __init__(self,nu0,width_v):
+    def __init__(self, nu0, width_v):
         self.nu0 = nu0
         self.width_v = width_v
-        self.width_nu = helpers.Delta_nu(Delta_v=self.width_v,nu0=self.nu0)
+        self.width_nu = helpers.Delta_nu(Delta_v=self.width_v, nu0=self.nu0)
         self.initialise_phi_nu_params()
         self.set_nu_extent()
         self.phi_nu0 = self.phi_nu(nu=self.nu0)
 
     def initialise_phi_nu_params(self):
         raise NotImplementedError
-    
-    def phi_nu(self,nu):
-        r'''The value of the line profile in frequency space.
-        
+
+    def phi_nu(self, nu):
+        r"""The value of the line profile in frequency space.
+
         The line profile is normalised such that its integral over frequency equals 1.
 
         Args:
             nu (float or numpy.ndarray): frequency in [Hz]
-        
+
         Returns:
             float or numpy.ndarray: The line profile in [Hz\ :sup:`-1`]
-        '''
+        """
         raise NotImplementedError
 
-    def phi_v(self,v):
-        r'''The value of the line profile in velocity space.
-        
+    def phi_v(self, v):
+        r"""The value of the line profile in velocity space.
+
         The line profile is normalised such that its integral over velocity equals 1.
 
         Args:
             v (float or numpy.ndarray): velocity in [m/s]
-        
+
         Returns:
             float or numpy.ndarra: The line profile in [(m/s)\ :sup:`-1`]
-        '''
-        nu = self.nu0*(1-v/constants.c)
-        return self.phi_nu(nu)*self.nu0/constants.c
+        """
[…]

[…]
