### 2026-01-27 #9233 @andresmegias
Hi all,

Thank you very much, @gica3618, for the updates and changes and for the explanations. I comment on them below:

The new comparison figure from the paper looks great! It's fantastic to see now the excellent agreement between Radex and pythonradex. I only a have an aesthetic suggestion, I think the axis labels would look better being centered, this can be done with `plt.supxlabel` and `plt.supylabel`.

Regarding the line profiles and the correction factor applied by Radex:

> I think for optical depth it is always correct, but for flux only if the line is optically thin.

Yes, makes sense.

Regarding the radius of the sphere with the spherical geometries, okay, I think I understand it now. The derivation of the flux density and escape probability by Osterbrock from 1974 (used by pythonradex for deriving the escape probability and specific intensity, and by Radex just for the escape probability) integrates all the flux from the whole sphere, therefore it implies that we do not resolve it, as you comment for pythonradex; the radius is actually implicit in the optical depth used in the formula (_τ_ = 2 _Rα_), along the sphere diameter. However, Radex is then focusing on the center of the sphere, so it assumes that we are resolving it. If I understood it correctly, this would be in contradiction with the derivation of the scape probability, since it does not depend on which part of the sphere we are looking at, so it would be like an average escape probability over all the sphere. With this reasoning I understand that your approach is indeed correct.

Then, the static spheric geometry assumes that we do not resolve the sphere but that it fills all the beam of our observations, otherwise we would have to apply a beam filling factor. Well, this is something that happens in general for both Radex and pythonradex, but it can be easily solved by manually applying this factor to the Radex or pythonradex output; and if we use a slab geometry, we are probably assuming that the source size is larger than our beam.

Regarding the analytical solution for the escape probability in the static slab: okay, I was mistaken, I thought the Radex _slab_ geometry was static!! I see it's a LVG slab (this should be better stated in Radex, I think). It is surprising to me that Radex does not account for a static slab geometry, since when using LTE we are normally assuming a static slab geometry, so in order to compare both Radex and LTE we should have the same geometry. In practice, since for the static sphere Radex uses the slab formula for the specific intensity, both cases are not that far away, but the formula used for the escape probability still does not correspond to a static slab.

Going back to the formula itself of the escape probability, you give the reference of Elitzur et al. (1992), but again I can't find the book freely available on the internet. Could you please add a little derivation of the formula? No need for schemes, since it's a slab, I hope it's not much work to do. Or either refer to a book or paper that is freely available.

Finally, thank you for all the changes you did in the new documentation! It looks really good and clear. (Ah yes, indeed when I talked about _intensity_ I meant _specific intensity_.) I only have two more very minor suggestions (feel free to take them into account or not):

- In most of the cases you use _τ_<sub>_ν_</sub> for the optical depth. I personally prefer to remove the _ν_ subindex, since for flux density and specific intensity it means that they are specific variables with respect to frequency (and we can consider the corresponding variables with respect to wavelength using the _λ_ subindex). However, optical depth is not specific to frequency nor wavelength. (Anyway, I know that many authors still use the ν subindex just to remember that it depends on frequency; others use the criterion I follow, like: https://www.cv.nrao.edu/~sransom/web/Ch2.html; you also didn't put the subindex in the formulae for LVG geometries in _Source geometries_). By the way, similarly, the absorption coefficient (_α_) is not specific, but the emissivity (_j_<sub>ν</sub>) is (you mention these variables in the _Theory_ section of the documentation).
- In some cells of the notebooks in _Benchmark vs RADEX_, there is a warning message coming from RADEX that gets repeated a few times: `Note: The following floating-point exceptions are signalling: IEEE_INVALID_FLAG`. I've also come across this message when running Radex and, honestly, I still don't understand what it means, but in many cases I get it when I'm doing something wrong. Maybe you could add a little clarification about this message (especially for cell 10 of _Comparing the outputs of pythonradex and RADEX_ and cell 12 of _Performance comparison between RADEX and pythonradex_), just to avoid people getting worried about it. My guess is that this happens when the physical conditions are too low or high, that's why you get some of these messages when running the grid. Also, I spotted a typo in this last notebook, in the paragraph before cell 13, last sentence: it says _Noet_ instead of _Note_.

Therefore, to me the code and paper is ready to be published! As I mentioned, I would appreciate seeing the derivation of the escape probability formula for a static slab, but it doesn't have to be right now, so I'm checking the last point of the check list that I had unchecked.

Thank you again, Gianni, for all the work done and all the clarifications!
