### 2026-01-28 f039e06
notebook updates and division by zero warning avoidance
