+++ src/pythonradex/intensity.py
@@ -197,7 +197,8 @@ class IntensityCalculator:
             S_line = line.source_function(x1=x1, x2=x2)
             S_nu += self.tau_nu_lines[i] * S_line
         S_nu += self.S_dust(self.nu) * self.tau_dust_nu
-        S_tot = np.where(self.tau_nu_tot == 0, 0, S_nu / self.tau_nu_tot)
+        S_tot = np.divide(S_nu, self.tau_nu_tot, out=np.zeros_like(S_nu),
+                          where=self.tau_nu_tot != 0)
         return S_tot
 
     def specific_intensity_spectrum(self):
