### 2026-01-28 fed74e3
added more text to docs about flux for spherical geometry (RADEX vs pythonradex)
