+++ tests/test_intensity.py
@@ -49,17 +49,17 @@ class TestFastIntensities:
     test_tau_values = [1e-5, 1e-2, 0.5, 10, 100]
 
     @staticmethod
-    def calculate_nu_and_tau_nu(lp, nu0, width_v, peak_tau):
+    def calculate_nu_and_tau(lp, nu0, width_v, peak_tau):
         width_nu = width_v / constants.c * nu0
         if lp == "rectangular":
             nu = np.linspace(nu0 - 0.55 * width_nu, nu0 + 0.55 * width_nu, 10000)
-            tau_nu = np.where(np.abs(nu - nu0) < width_nu / 2, peak_tau, 0)
+            tau = np.where(np.abs(nu - nu0) < width_nu / 2, peak_tau, 0)
         elif lp == "Gaussian":
             nu = np.linspace(nu0 - 3 * width_nu, nu0 + 3 * width_nu, 10000)
             sigma_nu = helpers.FWHM2sigma(width_nu)
-            tau_nu = np.exp(-((nu - nu0) ** 2) / (2 * sigma_nu**2))
-            tau_nu *= peak_tau / np.max(tau_nu)
-        return nu, tau_nu
+            tau = np.exp(-((nu - nu0) ** 2) / (2 * sigma_nu**2))
+            tau *= peak_tau / np.max(tau)
+        return nu, tau
 
     def test_fast_intensities(self):
         for (geo_name, geo), lp in itertools.product(
@@ -76,10 +76,10 @@ class TestFastIntensities:
                 expected_intensity = []
                 for t in self.test_transitions:
                     S = helpers.B_nu(nu=mol.nu0[t], T=Tex[t])
-                    nu, tau_nu = self.calculate_nu_and_tau_nu(
+                    nu, tau = self.calculate_nu_and_tau(
                         lp=lp, nu0=mol.nu0[t], width_v=mol.width_v, peak_tau=test_tau
                     )
-                    intensity_kwargs = {"tau_nu": tau_nu, "source_function": S}
+                    intensity_kwargs = {"tau": tau, "source_function": S}
                     if is_LVG_sphere:
                         intensity_kwargs["nu"] = nu
                         intensity_kwargs["nu0"] = mol.nu0[t]
@@ -125,7 +125,7 @@ class TestFastIntensities:
         # add the edge points, otherwise the interpolation will mess things up...
         nu = np.append(nu, (test_line.nu0 - width_nu / 2, test_line.nu0 + width_nu / 2))
         nu.sort()
-        expected_tau_nu = test_line.tau_nu(N1=N1, N2=N2, nu=nu)
+        expected_tau = test_line.tau(N1=N1, N2=N2, nu=nu)
         geo_name = "static slab"
         geo = radiative_transfer.Source.geometries[geo_name]
         tau_nu0_individual_transitions = mol.get_tau_nu0_lines(
@@ -144,13 +144,13 @@ class TestFastIntensities:
         transitions = [
             line_index,
         ]
-        constructed_nu, constructed_tau_nu, constructed_source_function = (
+        constructed_nu, constructed_tau, constructed_source_function = (
             intensity_calculator.get_intensity_parameters_for_rectangular(
                 transitions=transitions
             )
         )
-        interp_tau_nu = np.interp(x=constructed_nu, xp=nu, fp=expected_tau_nu)
-        assert np.allclose(constructed_tau_nu, interp_tau_nu, atol=0, rtol=1e-3)
+        interp_tau = np.interp(x=constructed_nu, xp=nu, fp=expected_tau)
+        assert np.allclose(constructed_tau, interp_tau, atol=0, rtol=1e-3)
         Tex = mol.get_Tex(level_population=level_pop)[line_index]
[…]
