### 2026-01-28 22a903f
updates for manual tests and benchmarks
