+++ manual_tests_and_benchmarks/performance_tests/run_source.py
@@ -76,7 +76,7 @@ for i in range(niter):
     tot_time += end - start
     print(f"solve time: {end-start:.3g}")
     start = time.time()
-    source.fluxes_of_individual_transitions(solid_angle=1, transitions=None)
+    source.frequency_integrated_emission(output_type="flux",solid_angle=1, transitions=None)
     end = time.time()
     tot_time += end - start
     print(f"flux time: {end-start:.3g}")
