### 2026-01-29 bd986be
work on output shape of spectrum, flux, etc
