### 2026-01-29 276f98e
added derivation of static slab escape probability to the docs
