### 2026-01-29 6c9e577
removed subscript \nu where needed
