+++ docs/geometry.rst
@@ -3,7 +3,7 @@
 Source geometries
 ======================
 
-The calculations of the escape probability and the emerging flux depend on the adopted geometry. Below we give an overview of the geometries available in ``pythonradex``. For all geometries, we assume a homogeneous medium (i.e. constant number density). For convenience, the escape probabilities are expressed as functions of optical depth :math:`\tau_\nu` rather than the absorption coefficient :math:`\alpha_\nu`. The specific intensities are expressed using the source function :math:`S_\nu`. In absence of dust and overlapping lines, it is simply given by the Planck function evaluated at the excitation temperature of the transition: :math:`S_\nu=B_\nu(T_\mathrm{ex})`.
+The calculations of the escape probability and the emerging flux depend on the adopted geometry. Below we give an overview of the geometries available in ``pythonradex``. For all geometries, we assume a homogeneous medium (i.e. constant number density). For convenience, the escape probabilities are expressed as functions of optical depth :math:`\tau` rather than the absorption coefficient :math:`\alpha`. The specific intensities are expressed using the source function :math:`S_\nu`. In absence of dust and overlapping lines, it is simply given by the Planck function evaluated at the excitation temperature of the transition: :math:`S_\nu=B_\nu(T_\mathrm{ex})`.
 
 Note that ``RADEX`` always uses the formula for a slab geometry to calculate the specific intensity, regardless of the adopted geometry. This is incorrect for spherical geometries. See :doc:`difference_pythonradex_RADEX` for more details.
 
@@ -15,14 +15,14 @@ A homogeneous, static sphere. The escape probability is given in [Osterbrock74]_
 .. math::
     :name: eq:beta_static_sphere
 
-    \beta(\tau_\nu) = \frac{3}{2\tau_\nu}\left(1-\frac{2}{\tau_\nu^2}+\left(\frac{2}{\tau_\nu}+\frac{2}{\tau_\nu^2}\right) e^{-\tau_\nu}\right)
+    \beta(\tau) = \frac{3}{2\tau}\left(1-\frac{2}{\tau^2}+\left(\frac{2}{\tau}+\frac{2}{\tau^2}\right) e^{-\tau}\right)
 
-where :math:`\tau_\nu` is the optical depth of the diameter of the sphere. The observed specific intensity (in [W/m\ :sup:`2`/Hz/sr]) can also be calculated from [Osterbrock74]_ and is given by:
+where :math:`\tau` is the optical depth of the diameter of the sphere. The observed specific intensity (in [W/m\ :sup:`2`/Hz/sr]) can also be calculated from [Osterbrock74]_ and is given by:
 
 .. math::
     :name: eq:intensity_static_sphere
 
-    I_\nu = \frac{2S_\nu}{\tau_\nu^2}\left(\frac{\tau_\nu^2}{2}-1+(\tau_\nu+1)e^{-\tau_\nu}\right)
+    I_\nu = \frac{2S_\nu}{\tau^2}\left(\frac{\tau^2}{2}-1+(\tau+1)e^{-\tau}\right)
 
[…]
