### 2026-01-29 #9233 @gica3618
Thanks again @andresmegias for carefully checking everything.

> I think the axis labels would look better being centered, this can be done with plt.supxlabel and plt.supylabel

Done.

> The derivation of the flux density and escape probability by Osterbrock from 1974 (used by pythonradex for deriving the escape probability and specific intensity, and by Radex just for the escape probability) integrates all the flux from the whole sphere, therefore it implies that we do not resolve it

I think for the derivation one does not need to consider whether the source is resolved or not. The derivation just calculates the mean probability for a photon to escape the sphere.

> However, Radex is then focusing on the center of the sphere, so it assumes that we are resolving it.

This might not be intentional by the Radex developers. It just turns out that with the formula they use to calculate the brightness temperature, they get the brightness temperature towards the center of the sphere. The excitation temperature and optical depth calculated by Radex are correct though.

> Then, the static spheric geometry assumes that we do not resolve the sphere but that it fills all the beam of our observations, otherwise we would have to apply a beam filling factor.

As commented above, the excitation temperature and optical depth are the same between RADEX and pyhonradex. The difference is just how to convert those into a brightness temperature and total flux. For total flux, RADEX overestimates it (unless the sphere is optically thick), while pythonradex gives the correct answer (this is independent of resolved or unresolved). For brightness temperature, it varies over the surface of the sphere, but these codes can only give back a single number. Radex gives the brightness temperature towards the center of the sphere, while pythonradex returns "total flux divided by solid angle of the sphere".

Yes good point about the beam filling factor, I didn't think about that. So indeed, pythonradex will agree with observations in terms of brightness temperature if the source is unresolved, but fills the beam (or an appropriate beam-filling factor was applied).

>  It is surprising to me that Radex does not account for a static slab geometry, since when using LTE we are normally assuming a static slab geometry, so in order to compare both Radex and LTE we should have the same geometry.

Maybe here I mis-understand something, but I think in LTE you don't need to assume any geometry?

> Going back to the formula itself of the escape probability, you give the reference of Elitzur et al. (1992), but again I can't find the book freely available on the internet. Could you please add a little derivation of the formula? No need for schemes, since it's a slab, I hope it's not much work to do.

I added the derivation to the documentation, please have a look.

> I personally prefer to remove the ν subindex, since for flux density and specific intensity it means that they are specific variables with respect to frequency

I never realised that the subindex means "specific", I always thought it just means that something depends on v. It makes sense, I removed the subindex from the documentation, notebooks, paper and also in the code (renamed the `tau_nu` method to `tau`).

> In some cells of the notebooks in Benchmark vs RADEX, there is a warning message coming from RADEX that gets repeated a few times: Note: The following floating-point exceptions are signalling: IEEE_INVALID_FLAG. I've also come across this message when running Radex and, honestly, I still don't understand what it means, but in many cases I get it when I'm doing something wrong. Maybe you could add a little clarification about this message

I also don't know where it comes from. It's a very good idea to add a note, otherwise people might think something is wrong. I added notes to both notebooks.

> Also, I spotted a typo in this last notebook, in the paragraph before cell 13, last sentence: it says Noet instead of Note.

Thank you, fixed!
