### 2026-01-29 #9233 @gica3618
I think for now I addressed all questions, so I will wait for further feedback.

If anyone wants to try the updated code or notebooks, please install pythonradex in editable mode from the joss_paper branch:

1. git clone -b joss_paper https://github.com/gica3618/pythonradex.git
2. cd pythonradex
3. pip install -e .

If you want to run a notebook, please download it from the joss_branch at https://github.com/gica3618/pythonradex/tree/joss_paper/docs/notebooks

Also, if you want to check the documentation, please visit the one corresponding to the joss_paper branch at https://pythonradex.readthedocs.io/en/joss_paper/
