### 2026-01-30 5b05faf
paper and docs updates
