+++ joss_paper/paper.md
@@ -70,13 +70,13 @@ The python programming language is now very widely used in astronomy. Still, no 
 
 ## Flux for spherical geometry
 
-**To calculate line fluxes, `pythonradex` uses different formulae depending on the geometry (slab or sphere; see the [`pythonradex` documentation](https://pythonradex.readthedocs.io/en/latest/index.html) for more details). On the other hand, `RADEX` always uses the formula for a slab. \autoref{fig:flux_comparison_sphere} illustrates the consequences. For a static sphere, `RADEX` overestimates the flux by a factor 1.5[^2] in the optically thin limit. The optically thin flux can easily be computed directly[^3], confirming that the flux computed by `pythonradex` is correct. In the optically thick case, only the surface of the static sphere is visible, so both codes agree despite using different formulae. On the other hand, for the LVG sphere, the difference is always a factor 1.5 regardless of optical depth. This is a consequence of LVG assumption that photons always escape unless absorbed locally.**
+**To calculate line fluxes, `pythonradex` uses different formulae depending on the geometry (slab or sphere; see the [`pythonradex` documentation](https://pythonradex.readthedocs.io/en/latest/index.html) for more details). On the other hand, `RADEX` always uses the formula for a slab. \autoref{fig:flux_comparison_sphere} illustrates the consequences. For a static sphere, by using the slab formula, the flux is overestimated by a factor 1.5[^2] in the optically thin limit. In the optically thick case, only the surface of the static sphere is visible, so the different formulae give the same result. On the other hand, for the LVG sphere, the difference is always a factor 1.5 regardless of optical depth. This is a consequence of the LVG assumption that photons always escape unless absorbed locally. By computing the optically thin flux directly[^3], it can be confirmed that the formulae used by `pythonradex` are correct.**
 
 [^2]: **The factor 1.5 simply corresponds to the volume ratio of a "spherical slab" (i.e. a cylinder) to a sphere.**
 [^3]: **In units of W/m$^2$, $F_\mathrm{thin} = V_\mathrm{sphere}n_2A_{21}\Delta E \frac{1}{4\pi d^2}$ with $V_\mathrm{sphere}=\frac{4}{3}R^3\pi$ the volume of the sphere, $n$ the number density, $x_2$ the fractional level population of the upper level, $A_{21}$ the Einstein coefficient, $\Delta E$ the energy of the transition and $d$ the distance.**
 
-<!-- This figure comes from  compare_emerging_flux_formula.py -->
-![Flux ratio between `RADEX` and `pythonradex` for spherical geometries as a function of optical depth.\label{fig:flux_comparison_sphere}](flux_comparison_spherical_geometries.pdf)
+<!-- This figure comes from  manual_test_and_benchmarks/emerging_flux/compare_emerging_flux_formula.py -->
+![For the two spherical geometries, we computed the flux using the slab formula (as done by `RADEX`) and a formula appropriate for a sphere (as done by `pythonradex`). The figure shows the ratio of the fluxes as function of optical depth.\label{fig:flux_comparison_sphere}](code/flux_comparison_spherical_geometries.pdf)
 
 
 # Dependencies
