+++ docs/images/readme_static_sphere_sketch.txt
@@ -1,1 +1,0 @@
-when exporting to jpg, increase the size to 40x40 inches2
