### 2026-01-30 f71a1bb
paper and docs updates
