### 2026-01-30 feb6acd
paper and docs updates
