### 2026-02-03 #9233 @andresmegias
Dear @gica3618,

Thank you very much for the clarifications. I've been reflecting about these topics of geometry and the derivation of the escape probability and specific intensity and there are still some things that I would like to comment and ask.

> I think for the derivation [of the escape probability] one does not need to consider whether the source is resolved or not. The derivation just calculates the mean probability for a photon to escape the sphere.

This is true if the telescope beam is larger than the source size, but if we are resolving the source and observing at a specific point, I think the escape probability can differ along the volume of the source. In a static sphere, for example, a photon close to the surface has more probability of escaping outside than one located in the center, right? If we observed the limb of the sphere, apart from using a formula for _I_<sub>_ν_</sub>(_τ_) that  integrates along the line of sight with our telescope beam (if it is very small, just _I_<sub>ν</sub>(_τ_) = _S_<sub>ν</sub>·(1 - e<sup>-_τ_</sup>)), I think one should use this local escape probability to model the level populations and compute the emission and absorption coefficients. I found a paper where they consider these variations in the escape probability for resolved sources: https://ui.adsabs.harvard.edu/abs/1992ApJ...401..416K/abstract.

> For total flux, RADEX overestimates it (unless the sphere is optically thick), while pythonradex gives the correct answer (this is independent of resolved or unresolved).

Well, as I commented before, if we observed a portion of the sphere with our telescope beam, even pythonradex would fail.

In any case, I don't think the scope of RADEX/pythonradex is observing a resolved sphere, so it is enough with having a clarification of this in the documentation, which I saw you already did regarding the intensity for the static sphere. I would only add that it's not only the intensity that changes, but the escape probability as well. Therefore, when using pythonradex/RADEX with a spheric geometry, this sphere should be unresolved (apparent size smaller than the telescope's beam size).

Now, regarding the beam filling factor, there's another subtlety to be taken into account that I didn't notice. If the sphere _just_ fills the beam of the telescope, we cannot use this source-averaged intensity (_I_<sub>_ν_</sub> = _F_<sub>_ν_</sub> / _Ω_) due to the antenna pattern; i.e., the intensity coming from the centre of the sphere would have a larger weight than the one coming from the borders in the resulting measured antenna temperature. Therefore, if the sphere is not optically thick, the perceived intensity will be a bit larger than the source-averaged one. If the beam is quite larger than the sphere, these differences would become insignificant, and just applying the beam filling factor would be exact. This is again to say that in RADEX/pythonradex, if we use a spheric geometry we should consider that the sphere is largely contained in our beam (so it is unresolved, and preferably, not very close to being resolved). This is contrast with the slab geometries, in which one assumes that the slab is quite larger than the beam size, if I'm not mistaken. Actually, in the extreme case of a very large sphere compared to our beam, the portion of the sphere that we observe will be mostly flat, like a slab.

I also wanted to point out a thing about the LVG sphere. You say that RADEX uses an incorrect formula for the specific intensity, since it misses the (1 - _v_<sup>2</sup>/_V_<sup>2</sup>) factor. Maybe I'm misunderstanding something but, if we focus on the peak intensity, wouldn't this mean _v_ = 0? If so, the _I_<sub>_ν_</sub>(_τ_) = _S_<sub>ν</sub>·(1 - e<sup>-_τ_</sup>) formula (which RADEX uses) would be correct.

> Maybe here I mis-understand something, but I think in LTE you don't need to assume any geometry?

I think so. We need to assume a geometry when integrating the radiative transfer equation to compute the expression of _I_<sub>_ν_</sub>(_τ_) to compare with observations. It's just that we normally use (in my experience) the typical formula of _I_<sub>_ν_</sub>(_τ_) = _S_<sub>ν</sub>·(1 - e<sup>-_τ_</sup>). This means that we are either in a slab (static of LVG) or in a LVG sphere, but not in a static sphere. Let me know if I'm wrong, my understanding is that in the end it's just like in RADEX/pythonradex, but we compute the emission and absorption coefficients (and then _τ_) assuming LTE (with the given column density and excitation temperature), without having to use the escape probability and the collisional coefficients.

Finally, I have a question regarding the values of the optical depth and column density that RADEX and pythonradex use in the case of the static sphere. In the corresponding formula for the specific intensity, _τ_ corresponds to the maximum optical depth, towards the centre of the sphere. Is this the value given by RADEX/pythonradex as an output? This also made me think: in the case of the column density given as an input, would this be an average column density over the source or the value at the centre? I ask because in order to compare with a slab, I think we should use an average column density. Similarly, I wonder if we should compare the average _τ_ in the sphere with the _τ_ in the slab.

> I added the derivation [for the static slab] to the documentation, please have a look.

Thank you very much! I noticed that for the LVG slab, the full reference of [Scoville74] has the wrong year, it says 1991 instead of 1974. Another thing I noticed is that in this reference the authors don't really derive the expression, they just give it. In the reference given in the RADEX paper (de Jong et al., 1975) they give more explanations, although it is somewhat obfuscated. Then you could also add that other reference in the documentation.

> I never realised that the subindex means "specific", I always thought it just means that something depends on v. It makes sense, I removed the subindex from the documentation, notebooks, paper and also in the code (renamed the tau_nu method to tau).

Nice!
