### 2026-02-04 9f7c383
making definition of column density and optical depth for spheres clear
