+++ src/pythonradex/radiative_transfer.py
@@ -23,7 +23,8 @@ class Source:
             containing all the information about the emitting atom or molecule
         tau_nu0_individual_transitions (numpy.ndarray): optical depth of each individual
             transition at the central frequency (contributions of dust or overlapping lines
-            to the optical depth are not included)
+            to the optical depth are not included). For spherical geometries,
+            the optical depth is along the diameter of the sphere.
         level_pop (numpy.ndarray): fractional population of each level
         Tex (numpy.ndarray): excitation temperature of each transition
         lower_level_population (numpy.ndarray): for each transition, the fractional
@@ -215,7 +216,9 @@ class Source:
                 a constant value for all frequencies. Defaults to None
                 (i.e. do not update).
             N (:obj:`float` or None): The column density in [m\ :sup:`-2`].
-                Defaults to None (i.e. do not update).
+                Defaults to None (i.e. do not update). For spherical geometries,
+                this corresponds to the column density along the diameter of the
+                sphere.
             Tkin (:obj:`float` or None): The kinetic temperature of the gas in [K].
                 Defaults to None (i.e. do not update).
             collider_densities (dict or None): A dictionary of the number densities of
@@ -233,8 +236,9 @@ class Source:
                 frequency. A single number is interpreted as a constant optical
                 depth for all frequencies. Can only be used with static geometries
                 (i.e. not with LVG geometries). Defaults to None (i.e. do not update).
-                For a model without dust, put this parameter to 0.
-
+                For a model without dust, put this parameter to 0. For a static
+                sphere, this corresponds to the optical depth along the diameter
+                of the sphere.
         """
         # why do I put this into a seperate method rather than __init__? The reason
         # is that the stuff in __init__ is expensive (in particular reading the
@@ -535,7 +539,8 @@ class Source:
 
     def tau(self, nu):
         r"""Calculate the total optical depth (all lines plus dust) at each
-        input frequencies
+        input frequencies. For spherical geometries, the returned optical depth
+        corresponds to the diameter of the sphere.
 
         Args:
             nu (0- or 1-dimensional array_like): The frequencies in [Hz]
@@ -668,76 +673,6 @@ class Source:
                 output_type=output_type,
             )
         return out.reshape(transitions_shape)
-
-    # def efficient_parameter_iterator(self,ext_backgrounds,N_values,Tkin_values,
-    #                                  collider_densities_values,T_dust=0,
-    #                                  tau_dust=0):
-    #     r'''Iterator to update parameters in an efficient way.
-    #         For all possible combinations of the input parameters
-    #         ext_backgrounds, N_values, Tkin_values and collider_densities_values,
-    #         the parameters are updated in each iteration.
-
-    #     Args:
-    #         ext_backgrounds (:obj:`dict`): A dictionary, one entry for
-    #             each background that should be used. Each entry can be a function
-    #             of frequency, or a single number (interpreted as a radiation field
-    #             independent of frequency). The units are W/m\ :sup:2/Hz/sr. The keys
-    #             of the dictionary are used in the output to identify which
-    #             background was used.
-    #         N_values (:obj:`list` or numpy.ndarray): The list of column densities to
-    #             compute models for, in [m\ :sup:`-2`].
-    #         Tkin_values (:obj:`list` or numpy.ndarray): The list of kinetic temperatures
-    #             to compute models for, in [K].
-    #         collider_densities_values (:obj:`dict`): A dictionary with one entry for each
-    #             collider. Each entry is a list of densities for which models should be
-    #             computed for, using a "zip" logic (i.e. calculate a model for the first
[…]
