### 2026-02-04 #9233 @gica3618
Hi @andresmegias, thanks again for carefully thinking about these problems. Here are my comments:

> > I think for the derivation [of the escape probability] one does not need to consider whether the source is resolved or not. The derivation just calculates the mean probability for a photon to escape the sphere.

> This is true if the telescope beam is larger than the source size, but if we are resolving the source and observing at a specific point, I think the escape probability can differ along the volume of the source. In a static sphere, for example, a photon close to the surface has more probability of escaping outside than one located in the center, right?

I was just talking about the calculation of the _mean_ escape probability by RADEX and pythonradex. Both codes calculate this mean escape probability and assume it applies everywhere. As you point out, this is not true: in reality, the escape probability is spatially dependent (same for a slab, of course). However, note that this spatial dependence is there whether you resolve the source or not. So even if you don't resolve the source, this simplification will introduce a certain amount of uncertainty (if you would do a proper 3D model, the total flux observed with unresolved observations might be different than what RADEX/pythonradex predict). Although I can see that if the source is resolved and you want to model different parts of the source, it becomes more important to take the spatial variations into account.

>>For total flux, RADEX overestimates it (unless the sphere is optically thick), while pythonradex gives the correct answer (this is independent of resolved or unresolved).

>Well, as I commented before, if we observed a portion of the sphere with our telescope beam, even pythonradex would fail.

Here I was talking about the total flux [W/m2] emitted by the sphere, not the flux from a specific localised region of the sphere. To summarise the spherical case:

- pythonradex gives the correct total flux [W/m2] emitted by the whole sphere, RADEX does not (unless the sphere is completely optically thick)
- as for brightness temperature, neither code gives the "correct" answer, since the brightness temperature is dependent on the line of sight, but the codes only give back a single number, so in fact they cannot give the "correct" answer. RADEX gives the brightness temperature towards the center of the sphere, while pythonradex gives "total flux divided by solid angle of the sphere".

If we talk about the flux from a specific localised region of the sphere, then yes, you probably want a more sophisticated model.

> In any case, I don't think the scope of RADEX/pythonradex is observing a resolved sphere, so it is enough with having a clarification of this in the documentation, which I saw you already did regarding the intensity for the static sphere. I would only add that it's not only the intensity that changes, but the escape probability as well. Therefore, when using pythonradex/RADEX with a spheric geometry, this sphere should be unresolved (apparent size smaller than the telescope's beam size).

As commented above, even if the source is unresolved, the escape probability is still spatially dependent in reality. This is just a limitation of the model. And this limitation also applies for a slab. On the other hand, I think you _can_ model resolved observations with pythonradex, in the sense that if you integrate over your image to calculate the total observed flux emitted by the sphere, you can compare that number with a prediction from a pythonradex model.

>Now, regarding the beam filling factor, there's another subtlety to be taken into account that I didn't notice. If the sphere just fills the beam of the telescope, we cannot use this source-averaged intensity (Iν = Fν / Ω) due to the antenna pattern; i.e., the intensity coming from the centre of the sphere would have a larger weight than the one coming from the borders in the resulting measured antenna temperature. Therefore, if the sphere is not optically thick, the perceived intensity will be a bit larger than the source-averaged one. If the beam is quite larger than the sphere, these differences would become insignificant, and just applying the beam filling factor would be exact. This is again to say that in RADEX/pythonradex, if we use a spheric geometry we should consider that the sphere is largely contained in our beam (so it is unresolved, and preferably, not very close to being resolved). This is contrast with the slab geometries, in which one assumes that the slab is quite larger than the beam size, if I'm not mistaken. Actually, in the extreme case of a very large sphere compared to our beam, the portion of the sphere that we observe will be mostly flat, like a slab.

This is a subtle point, but I think you are right; in terms of brightness temperature, pythonradex will give the right answer if the source is completely unresolved, and the right beam-filling factor is applied. I adjusted the documentation.

As commented above though, you can still model resolved observations in terms of the total flux by integrating over your image.

As for the slab, I don't think we need to assume that it is large compared to the beam. Could you elaborate why you think this is the case?

> I also wanted to point out a thing about the LVG sphere. You say that RADEX uses an incorrect formula for the specific intensity, since it misses the (1 - v2/V2) factor. Maybe I'm misunderstanding something but, if we focus on the peak intensity, wouldn't this mean v = 0? If so, the Iν(τ) = Sν·(1 - e-τ) formula (which RADEX uses) would be correct.

Yes, for the peak intensity, both codes give the same answer. The difference is the frequency dependence of the specific intensity, and as a consequence also the total flux (i.e. the specific intensity integrated over frequency and solid angle).

> >Maybe here I mis-understand something, but I think in LTE you don't need to assume any geometry?

> I think so. We need to assume a geometry when integrating the radiative transfer equation to compute the expression of Iν(τ) to compare with observations. It's just that we normally use (in my experience) the typical formula of Iν(τ) = Sν·(1 - e-τ). This means that we are either in a slab (static of LVG) or in a LVG sphere, but not in a static sphere. Let me know if I'm wrong, my understanding is that in the end it's just like in RADEX/pythonradex, but we compute the emission and absorption coefficients (and then τ) assuming LTE (with the given column density and excitation temperature), without having to use the escape probability and the collisional coefficients.

Yes, to calculate the intensity and flux, you need to assume a geometry even for LTE. I was just thinking about the level populations. But going back to where this discussion started, namely that RADEX does not offer a static slab, and how we can compare RADEX slab results to LTE, I think an LTE static slab and an LTE LVG slab are very similar, the difference just being the velocity structure.

> Finally, I have a question regarding the values of the optical depth and column density that RADEX and pythonradex use in the case of the static sphere. In the corresponding formula for the specific intensity, τ corresponds to the maximum optical depth, towards the centre of the sphere. Is this the value given by RADEX/pythonradex as an output? This also made me think: in the case of the column density given as an input, would this be an average column density over the source or the value at the centre? I ask because in order to compare with a slab, I think we should use an average column density. Similarly, I wonder if we should compare the average τ in the sphere with the τ in the slab.

Both the input column density and the output optical depth are along the diameter of the sphere. This is an important point, and I made this point clear now in the docstrings and in the example notebooks.

> I noticed that for the LVG slab, the full reference of [Scoville74] has the wrong year, it says 1991 instead of 1974. Another thing I noticed is that in this reference the authors don't really derive the expression, they just give it. In the reference given in the RADEX paper (de Jong et al., 1975) they give more explanations, although it is somewhat obfuscated. Then you could also add that other reference in the documentation.

The year is fixed. I also added the reference to de Jong+75, although I agree that even they do not really derive the formula.
