### 2026-02-04 #9233 @andresmegias
Hi again, @gica3618,

Thank you very much for the explanations about the geometries and the calculations of the escape probabilities and intensities. Okay, I understand, I didn't realise that even in an unresolved spherical source, one would have to take into account the local escape probability to compute the total flux emitted from the sphere.

> This is a subtle point, but I think you are right; in terms of brightness temperature, pythonradex will give the right answer if the source is completely unresolved, and the right beam-filling factor is applied.

Just to be sure, here by _right answer_ you mean ignoring this little "error" coming from using a mean escape probability for all the volume that you commented before?

> On the other hand, I think you can model resolved observations with pythonradex, in the sense that if you integrate over your image to calculate the total observed flux emitted by the sphere, you can compare that number with a prediction from a pythonradex model.

Hum, but you would only integrate the part of the sphere that lies in your telescope beam, so well, it could serve to infer the size of the sphere, did you mean that? Not very import, just want to understand it.

> As for the slab, I don't think we need to assume that it is large compared to the beam. Could you elaborate why you think this is the case?

Hm, no, it's true that it could also be smaller than the beam, in which case one would apply a beam filling factor. I was thinking that is more usual in my experience, e.g. to model extense sources. In a slab, the expression for _I_<sub>_ν_</sub>(_τ_) is the same whether resolving or not the source, so it doesn't matter that it's smaller than the beam (as long as we use the correct filling factor).

> Yes, for the peak intensity, both codes give the same answer. The difference is the frequency dependence of the specific intensity, and as a consequence also the total flux (i.e. the specific intensity integrated over frequency and solid angle).

Okay, then RADEX is giving an incorrect flux but a proper peak Rayleigh-Jeans brightness temperature (specific intensity), so I would clarify this (shortly) in the documentation. The case of the static sphere is clearly worse because it is failing even in the peak, so the Rayleigh-Jeans temperature given by RADEX is also wrong.

> Yes, to calculate the intensity and flux, you need to assume a geometry even for LTE. I was just thinking about the level populations. But going back to where this discussion started, namely that RADEX does not offer a static slab, and how we can compare RADEX slab results to LTE, I think an LTE static slab and an LTE LVG slab are very similar, the difference just being the velocity structure.

Okay, I agree.

> Both the input column density and the output optical depth are along the diameter of the sphere. This is an important point, and I made this point clear now in the docstrings and in the example notebooks.

This is really nice to know!! Then, in order to compare geometries, it would be more convenient to use the average optical depth towards the sphere, which is 2/3 of the maximum, right? Interestingly, if we introduce this scaling in the formula for the specific intensity in a static sphere, it now gets so much closer to the slab cases:

In fact, the 1.5 (3/2) factor of difference at optically thin lines disappears. The ratio of slab / static sphere starts being 0 at _τ_ = 0, it gets a maximum of ~1.05 at _τ_ = 2, and then decreases at a slower rate, being ~1.01 at _τ_ = 10:

For the case of spheric geometries, this also seems more intuitive, since before the escape probabilities with respect to optical depth of the static sphere and LVG sphere were somewhat off, but now they lie quite close:

(I also attach the little script I used to generate the plots: [geometryplots.py](https://github.com/user-attachments/files/25080238/geometryplots.py))

In any case, this does not change that RADEX is using the wrong formula, since it considers the optical depth at the centre of the sphere.

Okay, I think I understand now why no one has noticed this wrong formula. I think that normally people compare RADEX (with the default geometry of the static sphere) with LTE (which normally assumes a slab). For optically thin lines, on the one hand RADEX is overestimating the intensity by a factor of 3/2, so if we do a a fit we underestimate the column density by a factor of 2/3; but on the other hand, this output column density is the one at the centre, which is 3/2 of the average one (the one equivalent to the slab). Therefore, these two factors cancel and we get very similar column densities.
