### 2026-02-06 #9233 @gica3618
Hi @andresmegias!

> >This is a subtle point, but I think you are right; in terms of brightness temperature, pythonradex will give the right answer if the source is completely unresolved, and the right beam-filling factor is applied.

> Just to be sure, here by right answer you mean ignoring this little "error" coming from using a mean escape probability for all the volume that you commented before?

Yes exactly, I mean the right answer in the framework of the model where there is no spatial variation of the radiation field, level populations etc.

> >On the other hand, I think you can model resolved observations with pythonradex, in the sense that if you integrate over your image to calculate the total observed flux emitted by the sphere, you can compare that number with a prediction from a pythonradex model.

> Hum, but you would only integrate the part of the sphere that lies in your telescope beam, so well, it could serve to infer the size of the sphere, did you mean that? Not very import, just want to understand it.

What I mean is if you have a resolved image of the source (from interferometry, or by doing several pointings with a single dish telescope). I mean the whole sphere appears in your image (not just parts of it) and is resolved. Then you can derive the total flux (W/m2) emitted by the sphere by integrating over the source in the image. Plus you know the size of the sphere from your image. Then you can compare the total flux to a pythonradex model.

> Okay, then RADEX is giving an incorrect flux but a proper peak Rayleigh-Jeans brightness temperature (specific intensity), so I would clarify this (shortly) in the documentation. The case of the static sphere is clearly worse because it is failing even in the peak, so the Rayleigh-Jeans temperature given by RADEX is also wrong.

Yes I think that's correct. I adjusted the documentation. Note that in terms of total flux, LVG sphere is actually worse than the static sphere, since it is always off by a factor 1.5 (while the static sphere is only off in the optically thin case).

> This is really nice to know!! Then, in order to compare geometries, it would be more convenient to use the average optical depth towards the sphere, which is 2/3 of the maximum, right? Interestingly, if we introduce this scaling in the formula for the specific intensity in a static sphere, it now gets so much closer to the slab cases:

> In fact, the 1.5 (3/2) factor of difference at optically thin lines disappears. The ratio of slab / static sphere starts being 0 at τ = 0, it gets a maximum of ~1.05 at τ = 2, and then decreases at a slower rate, being ~1.01 at τ = 10:

> For the case of spheric geometries, this also seems more intuitive, since before the escape probabilities with respect to optical depth of the static sphere and LVG sphere were somewhat off, but now they lie quite close:

Thanks for the plots! When I calculate the mean optical depth, I get 0.78 of the maximum. But anyway, I think the results seen in your first two figures can be explained as follows:

- In the optically thin limit, the intensity becomes proportional to the optical depth (for both slab formula and the static sphere formula). So by multiplying the optical depth by 3/2, you are effectively multiplying the intensity by 3/2. And since the intensity ratio between RADEX and pythonradex is exactly 3/2 in the optically thin limit, you get perfect agreement. (as explained in the docs and the paper, the intensity or flux ratio of 3/2 is just the volume ratio of a cylinder and a sphere; by using the slab formula, RADEX effectively transforms the sphere into a "circular slab", i.e. a cylinder).
- In the optically thick limit, the intensity is independent of tau, and just becomes equal to the source function (again for both formulae), so multiplying the optical depth by 3/2 has no effect and again you get perfect agreement.

> Okay, I think I understand now why no one has noticed this wrong formula. I think that normally people compare RADEX (with the default geometry of the static sphere) with LTE (which normally assumes a slab). For optically thin lines, on the one hand RADEX is overestimating the intensity by a factor of 3/2, so if we do a a fit we underestimate the column density by a factor of 2/3; but on the other hand, this output column density is the one at the centre, which is 3/2 of the average one (the one equivalent to the slab). Therefore, these two factors cancel and we get very similar column densities.

By comparing to LTE, do you mean that you run RADEX with a high collider density such that LTE applies? But yes, what you say might make sense. Another way to say it is that effectively, RADEX is considering a "circular slab", so if LTE applies, you will get the same result as from an LTE slab model.
