### 2026-02-06 #9233 @andresmegias
Hi, @gica3618,

> Yes exactly, I mean the right answer in the framework of the model where there is no spatial variation of the radiation field, level populations etc.

Okay, thanks!

> What I mean is if you have a resolved image of the source (from interferometry, or by doing several pointings with a single dish telescope). I mean the whole sphere appears in your image (not just parts of it) and is resolved. Then you can derive the total flux (W/m2) emitted by the sphere by integrating over the source in the image. Plus you know the size of the sphere from your image. Then you can compare the total flux to a pythonradex model.

Sure, that's totally right!

> Note that in terms of total flux, LVG sphere is actually worse than the static sphere, since it is always off by a factor 1.5 (while the static sphere is only off in the optically thin case).

Ahh, yes, because of the integration of the parabolic profile, no? Which is in the end the integral from 0 to 1 of 1 - *x*<sup>2</sup>, so 1 - 1/3 = 2/3.

By the way, you say in the documentation that the current formulae used by RADEX for the LVG sphere escape probability are different than the simple *β*(*τ*) = (1 - exp(-*τ*)) / *τ*. The one used for *τ* > 7 is indeed mentioned in the RADEX manual (https://var.sron.nl/radex/radex_manual.pdf, section 3.2) as convenient for a turbulent medium. I still don't see any explanation of why it's used for *τ* > 7 in the LVG sphere. Then I see that the formula used for *τ* < 7 is in the end the same of (1 - exp(-*τ*)) / *τ* but changing *τ* by 1.17 *τ* (I wonder why).

> Thanks for the plots! When I calculate the mean optical depth, I get 0.78 of the maximum.

I checked the maths and it should be 2/3 of the maximum. It's in the end the calculation of the average of the length of the chords of a sphere that are parallel to our line of sight (since we are very far away from the sphere), averaging over the apparent circle that we see.

If _r_ if the distance to the centre of the circle, and the length of the chord is *L*(*r*), we should integrate 2π*R* *L*(*r*) d*r* from 0 to *r*, and then divide by π*R*<sup>2</sup>. Using the same angle *θ* from the derivation of the formulae for the static sphere in the documentation, we can do the change *r*(*θ*) = *R* sin(*θ*), so *L*(*θ*)  =  2*R* cos(*θ*) and d*r* = *R* cos(θ) d*θ*. Then, the expression to integrate would be 4π*R*<sup>3</sup> cos<sup>2</sup>*θ* sin*θ* d*θ*, from 0 to π/2, which gives 4/3 π*R*<sup>3</sup>. Then, dividing by π*R*<sup>2</sup> we get 2/3 2R, so 2/3 of the maximum length (the diameter). This should apply also for the optical depth and the column density since we assume an homogeneous sphere.

> (as explained in the docs and the paper, the intensity or flux ratio of 3/2 is just the volume ratio of a cylinder and a sphere; by using the slab formula, RADEX effectively transforms the sphere into a "circular slab", i.e. a cylinder)

It's another way of seeing it, yes, a cylinder with the same surface area than the sphere and the length/depth equal to the diameter. However, then RADEX is using a wrong formula for the escape probability, which I assume is not the same as in a sphere. I guess that for optically thin lines, since the escape probabilities are always very high, there is not a lot of difference between using a sphere or a cylindric slab or the default thin slab. Still they would be two different cases, since one is very thin and the other not. So in the end the conclusion is that for low opacities the static sphere geometry in RADEX is actually equivalent to a slab.

(Edit: I think I remember why I thought the escape probability for the static (thin) slab can only be used if it is quite larger than the telescope beam. In the calculation of the flux density, the angle of incidence *θ* is integrated from 0 to π/2, so this flux at the surface of the slab applies only for the regions that are quite far away from the edges; for the regions that are closer to the edges, the maximum value of *θ* will depend on the other angle *φ*; in other words, the optical depth of the rays have a maximum value towards the edges. So in the end I interpret that the computed average escape probability applies as long as the regions we are observing are quite far away from the edge, which happens if the slab is quite larger than the telescope beam.)

> By comparing to LTE, do you mean that you run RADEX with a high collider density such that LTE applies?

Yes, that's it. Since the average optical depth in a sphere is 2/3 of the maximum one, the factors 2/3 and 3/2 exactly cancel for the case of optically thin lines.

Thank you very much, Gianni, for the useful discussions. After all them, I have all my doubts clear and from my side the paper can be published (I mean, I already said it a few comments ago, but now I understand everything better and I appreciate the clarifications very much). It is a really nice work and a very useful code, congratulations, Gianni!
