### 2026-02-06 #9233 @warrickball
> After all them, I have all my doubts clear and from my side the paper can be published (I mean, I already said it a few comments ago, but now I understand everything better and I appreciate the clarifications very much). It is a really nice work and a very useful code, congratulations, Gianni!

Thanks @andresmegias for the thorough review. I've been quietly following the conversation and appreciate the effort and detail.

With @GijsVermarien's [recommendation for publication], the paper has the two reviews we need and is in effect accepted. @pscicluna, thanks for your remarks, which certainly improved the submission anyway. We only tag complete reviews though, so I'll remove you from the metadata shortly.

@gica3618, congratulations! I'll start the post-review process in a moment by creating a checklist for me to work through (including some editorial checks) and some items for you. These will be for you to

* create a versioned release and let me know the version number, and
* archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or another similar service) and let me know the DOI.

Please check that the metadata (title, authors, licence) of the archive matches the JOSS paper. If a Zenodo archive is generated automatically (e.g. using the Zenodo-GitHub integration: [GitHub docs](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content), [Zenodo docs](https://help.zenodo.org/docs/github/)), the metadata can be mismatched for a number of reasons.

Changing the archive metadata usually doesn't require updating the files and we aren't worried about the paper, so you won't need a new archive after I've done my additional editorial checks.
