### 2026-02-06 #9233 @warrickball
The only immediate issue with the DOIs is

```
INVALID DOIs

- https://doi.org/10.1002/9783527618170.ch1 is INVALID because of 'https://doi.org/' prefix
```

This suggests there's a BibTeX entry where the `doi` field is a complete URL but should just be the DOI itself (i.e. `10.1002/9783527618170.ch1`).
