### 2026-02-10 #9233 @gica3618
Hi @andresmegias!

> >Note that in terms of total flux, LVG sphere is actually worse than the static sphere, since it is always off by a factor 1.5 (while the static sphere is only off in the optically thin case).

>Ahh, yes, because of the integration of the parabolic profile, no? Which is in the end the integral from 0 to 1 of 1 - x2, so 1 - 1/3 = 2/3.

The way I see it (which I think is equivalent to what you say) is that for the LVG sphere, you effectively are seeing a series of circular slabs (thin cylinders), each at a different velocity (see the derivation of the LVG sphere formulae in the documentation). The largest slab with a surface area of R^2*pi is going through the center of the sphere and has zero velocity. As you move away from the center along the line of sight, the slabs get smaller and smaller. However, even if the optical depth of each individual slab is very high, you still see them all, since they are all at different velocities. But by using the slab formula (like RADEX), you are effectively assuming that all slabs have the same surface area as the largest slab at v=0. So whether optically thin or thick, you will always overestimate the total flux by a factor 3/2.

> By the way, you say in the documentation that the current formulae used by RADEX for the LVG sphere escape probability are different than the simple β(τ) = (1 - exp(-τ)) / τ. The one used for τ > 7 is indeed mentioned in the RADEX manual (https://var.sron.nl/radex/radex_manual.pdf, section 3.2) as convenient for a turbulent medium. I still don't see any explanation of why it's used for τ > 7 in the LVG sphere.

I think it is because once tau becomes smaller than 2*sqrt(pi) (which evaluates to ~3.54), the formula for turbulent medium becomes undefined (square root from a negative number). So for smaller tau, a different formula has to be used. I am not why tau=7 was chosen for the transition to a different formula.

> Then I see that the formula used for τ < 7 is in the end the same of (1 - exp(-τ)) / τ but changing τ by 1.17 τ (I wonder why).

Interesting! I didn't notice that. But at least with this formula, at tau=7, both formulas give approximately the same value.

> I checked the maths and it should be 2/3 of the maximum. It's in the end the calculation of the average of the length of the chords of a sphere that are parallel to our line of sight (since we are very far away from the sphere), averaging over the apparent circle that we see.

You are right, it is indeed 2/3. My calculation was wrong.

> (Edit: I think I remember why I thought the escape probability for the static (thin) slab can only be used if it is quite larger than the telescope beam. In the calculation of the flux density, the angle of incidence θ is integrated from 0 to π/2, so this flux at the surface of the slab applies only for the regions that are quite far away from the edges; for the regions that are closer to the edges, the maximum value of θ will depend on the other angle φ; in other words, the optical depth of the rays have a maximum value towards the edges. So in the end I interpret that the computed average escape probability applies as long as the regions we are observing are quite far away from the edge, which happens if the slab is quite larger than the telescope beam.)

Yes, for the derivation of the escape probability and total flux, it is assumed that the slab is thin (i.e. the extent in x and y is much larger than the depth of the slab). What it means is that if you look at the slab as a whole, the edge effects are small (most of the emission comes from regions far away from the edges). So even if you don't resolve it, there is no problem. I guess the only issue would be if you resolve it and then specifically focus on the edge region.

I would also like to thank you for this very useful discussion which really improved the code, documentation and notebooks a lot!
