+++ joss_paper/paper.bib
@@ -25,7 +25,6 @@ title = {FUNDAMENTALS OF RADIATIVE TRANSFER},
 booktitle = {Radiative Processes in Astrophysics},
 chapter = {1},
 pages = {1-50},
-doi = {https://doi.org/10.1002/9783527618170.ch1},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/9783527618170.ch1},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/9783527618170.ch1},
 year = {1985},
