+++ joss_paper/paper.md
@@ -28,7 +28,7 @@ A common task in astronomical research is to estimate the physical parameters (t
 
 Modern astronomical facilities such as the Atacama Large Millimeter/submillimeter Array (ALMA) or the James Webb Space Telescope (JWST) are providing a wealth of line emission data at radio and infrared wavelengths. These data are crucial to constrain the physical and chemical properties of various astrophysical environments.
 
-To interpret such data, a radiative transfer calculation is typically used [see @Rybicki1985 for an introduction to radiative transfer]. For a given set of input parameters describing the source (temperature, density, geometry, etc.), one calculates the amount of radiation reaching the telescope. The input parameters are then adjusted such that the predicted flux matches the observations.
+To interpret such data, a radiative transfer calculation is typically used [see @Rybicki1985, for an introduction to radiative transfer]. For a given set of input parameters describing the source (temperature, density, geometry, etc.), one calculates the amount of radiation reaching the telescope. The input parameters are then adjusted such that the predicted flux matches the observations.
 
 If the medium is dense enough, local thermodynamic equilibrium (LTE) can be assumed. This considerably simplifies the radiative transfer calculation. A non-LTE calculation is considerably more complex and computationally expensive because the fractional population of the molecular energy levels needs to be solved for numerically [e.g. @vanderTak2007]. Typically, an iterative approach is used: from a first guess of the level populations, the radiation field is calculated. This radiation field is then used to solve for updated level populations. The iterations continue until convergence is reached.
 
