+++ joss_paper/paper.md
@@ -34,13 +34,13 @@ If the medium is dense enough, local thermodynamic equilibrium (LTE) can be assu
 
 Various codes are available to solve the radiative transfer. Codes solving the radiative transfer in 3D are used for detailed calculations of sources with well-known geometries. Examples include `RADMC-3D` [@Dullemond2012] and `LIME` [@Brinch2010]. However, a full 3D calculation is often too computationally expensive if a large parameter space needs to be explored, in particular in non-LTE. 1D codes that quickly provide an approximate solution are a commonly used alternative. In this respect, the 1D non-LTE code `RADEX` [@vanderTak2007] has gained considerable popularity: as of February 10, 2026, the paper presenting `RADEX` [@vanderTak2007] has 1463 citations. The Fortan code `RADEX` solves the radiative transfer of a uniform medium using an escape probability formalism.
 
-The python programming language is now widely used in astronomy. Still, no python version of `RADEX` is available, although some python wrappers (for example `SpectralRadex` [@SpectralRadex] or `ndradex` [@ndradex]) and even a Julia version [`Jadex`, @jadex] exist. Furthermore, `RADEX` cannot take into account the effects of an internal continuum field (typically arising from dust that is mixed with the gas), nor cross-excitation effects arising when transitions overlap in frequency. The `pythonradex` code addresses these concerns.
+The python programming language is now widely used in astronomy. Still, no python version of `RADEX` is available, although some python wrappers such as `SpectralRadex` [@SpectralRadex] or `ndradex` [@ndradex], and even a Julia version [`Jadex`, @jadex], exist. Furthermore, `RADEX` cannot take into account the effects of an internal continuum field (typically arising from dust that is mixed with the gas), nor cross-excitation effects arising when transitions overlap in frequency. The `pythonradex` code addresses these concerns.
 
 # Implementation
 
 `pythonradex` is written in python and implements the Accelerated Lambda Iteration (ALI) scheme presented by @Rybicki1992. Like `RADEX`, an escape probability equation is used to calculate the radiation field for a given level population. This allows solving the radiative transfer iteratively. To speed up the convergence, ng-acceleration [@Ng1974] is employed. To improve performance, critical parts of the code are just-in-time compiled using `numba` [@Lam2015].
 
-`pythonradex` supports four geometries: two static geometries (slab and sphere), and two large-velocity-gradient (LVG) geometries (again slab and sphere). In the LVG approximation, it is assumed that all regions of the source are Doppler shifted with respect to each other due to a velocity gradient. This means that all photons escape the source, unless absorbed locally (i.e. close to the emission location; e.g. @Elitzur1992).
+`pythonradex` supports four geometries: two static geometries (slab and sphere), and two large-velocity-gradient (LVG) geometries (again slab and sphere). In the LVG approximation, it is assumed that all regions of the source are Doppler shifted with respect to each other due to a velocity gradient. This means that all photons escape the source, unless absorbed locally [i.e. close to the emission location; e.g. @Elitzur1992].
 
 Currently, effects of internal continuum and overlapping lines can only be included for the static geometries. Another limitation is that only a single molecule can be considered at a time. Thus, solving the radiative transfer of overlapping lines of different molecules is not supported yet. Also, treating overlapping lines adds considerable computational cost because averages over line profiles need to be calculated.
 
