### 2026-02-10 #9233 @gica3618
@warrickball thank you very much for your support. I fixed the issue with the DOI and made some adjustments to the paper (such as removing bold text, fixing reference parenthesis etc,). One question, for the new release, is it ok if I merge the joss_paper branch into the main branch before the release? Or is it going to mess up things because this issue is linked to the joss_paper branch?
