### 2026-02-10 #9233 @warrickball
You can merge the `joss_paper` branch into `main` if you like. Just let me know and I'll update the metadata, which is straightforward. All that really matters is that the branch is correctly set at publication time so that our pipeline can find the correct `paper.md` and `paper.bib` source files.
