### 2026-02-17 a4de76d
added CITATION.cff; minor paper and readme updates
