+++ README.txt
@@ -1,7 +1,7 @@
-pythonradex solves the 1D non-LTE radiative transfer for a homogeneous medium. It is a tool to quickly estimate the radiation from an astrophysical gas. It also provides a convenient function to read data files from the LAMDA database. pythonradex is a python re-implementation of the RADEX code, with additional functionality not included in RADEX (treatment of overlapping lines, treatment of dust continuum, output of spectra,...)
+pythonradex solves the 1D non-LTE radiative transfer for a homogeneous medium. It is a tool to quickly estimate the radiation from an astrophysical gas. It also provides a convenient function to read data files in the LAMDA format. pythonradex is a python re-implementation of the RADEX code, with additional functionality not included in RADEX (treatment of overlapping lines, treatment of dust continuum, output of spectra, and more).
 
 For more details, please refer to the documentation: http://pythonradex.readthedocs.io/
 
 For questions, bug reports, or feature suggestions, please open an issue on github.
 
-pythonradex is written and maintained by Gianni Cataldi
+pythonradex is written and maintained by Gianni Cataldi.
