### 2026-02-17 #9233 @gica3618
Hi @warrickball, here are some updates:

- I merged the joss_paper branch into the main branch. So now the paper is in the main branch.
- I made a new release, v2.0.0 (https://github.com/gica3618/pythonradex/releases/tag/v2.0.0)
- The new release is archived on zenodo: https://doi.org/10.5281/zenodo.18666067
