### 2026-02-18 c5b3511
removed __about__.py
