+++ src/pythonradex/__about__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: 2025-present Gianni Cataldi <cataldi.gia@gmail.com>
-#
-# SPDX-License-Identifier: MIT
-__version__ = "0.0.1"
