### 2026-02-18 7155e6f
added __version__ to __init__.py
