+++ src/pythonradex/__init__.py
@@ -4,3 +4,6 @@ Python implementation of the RADEX radiative transfer code
 """
 
 from . import helpers, LAMDA_file, radiative_transfer
+from importlib.metadata import version
+
+__version__ = version("pythonradex")
