### 2026-02-24 #9233 @gica3618
Hi @warrickball, thanks for your edits! I merged your PR of initial editorial changes (https://github.com/gica3618/pythonradex/pull/24).
Also, I changed the branch to "development", to keep "master" at v2.0.0.
