### 2026-02-24 #9233 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
