### 2026-02-24 #9233 @warrickball
Thanks once again to @GijsVermarien and @andresmegias for reviewing this submission. JOSS simply wouldn't be possible without its community of volunteers. Thanks too to @pscicluna for the comments that still contributed to the submission, even if it didn't end up a full review.
