### 2026-02-25 #9233 @GijsVermarien
@gica3618 Congrats on getting the paper accepted :) Thank you @andresmegias for your thorough review and @warrickball for guiding me as first time reviewer for JOSS.
