### 2026-05-01 #9233 @gica3618
Dear @warrickball, dear reviewers, I realised that I might have done a mistake in the paper. The paper says "Still, no Python version of RADEX is available, although some Python wrappers such as ... exist." However, I realised that the molsim package (https://github.com/bmcguir2/molsim/) contains a RADEX module fully written in python, although I cannot find any documentation. So my statement that there is no python version available is not really correct. @warrickball could you advise how this should be handled?
